### 2021-06-30 #3436 @whedon
[PRE REVIEW]: PyVOLCANS: A Python package to flexibly explore similarities and differences between volcanic systems
**Submitting author:** @PTierz (<a href="http://orcid.org/0000-0001-8889-9900">Pablo Tierz Lopez</a>)
**Repository:** <a href="https://github.com/BritishGeologicalSurvey/pyvolcans" target ="_blank">https://github.com/BritishGeologicalSurvey/pyvolcans</a>
**Version:** v1.1.2
