### 2021-07-01 #3436 @elbeejay
Hello @PTierz, thanks for your submission to JOSS. I applaud you and your co-authors on porting your MATLAB code to Python in an effort to make VOLCANS fully open-access. I have a few questions during this pre-review stage of the JOSS process:

1. In the project `README.md` it is [mentioned](https://github.com/BritishGeologicalSurvey/pyvolcans#analogy-matrices) that there are plans to convert the calculation of analogy matrices to Python - what is the anticipated timeline for this addition? I think adding this component to the code base would make this submission stronger and the methodology more transparent, by making it easier for a user to modify and/or extend the code if needed.
2. Beyond the information provided in the project `README.md` I have been unable to find further documentation (e.g. API documentation, community guidelines). Can you direct me to where I might find this information? The expected documentation for JOSS submissions can be found [here](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation).

Thank you, let me know if any of these questions require further clarification.
