### 2021-07-02 #3436 @PTierz
Hello @elbeejay,

Very many thanks for your message and your encouraging words. We sincerely believe that `PyVOLCANS` [v1.1.2](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/v1.1.2) can already represent a very important step forward for our community.

Unfortunately, the second half of this week has been quite busy, and I will not be able to provide you with substantial feedback on your two queries by the end of today. Please accept my apologies.

I am currently in discussions with the co-authors to be able to send you appropriate, full responses to your questions over the following 1-2 weeks, OK?

Thank you very much in advance for your support.

Best regards,
Pablo
