### 2021-07-02 #3436 @elbeejay
> I am currently in discussions with the co-authors to be able to send you appropriate, full responses to your questions over the following 1-2 weeks, OK?

Great, this is no problem, please take as much time as you need.
