### 2021-07-16 1bb95cb
Extract detailed documentation to put on wiki page
