+++ README.md
@@ -1,53 +1,48 @@
 # PyVOLCANS
 
-An open-access Python tool that generates data-driven sets of analogue
-volcanoes for any Holocene volcano listed in the Global Volcanism Program (GVP)
-Volcanoes Of The World database (v. 4.6.7), based on the VOLCANS (VOLCano
-ANalogues Search) method presented by Tierz, Loughlin and Calder (2019):
-[https://doi.org/10.1007/s00445-019-1336-3](https://doi.org/10.1007/s00445-019-1336-3)
+> PyVOLCANS: A Python package to flexibly explore similarities and differences between volcanic systems
 
-VOLCANS uses five volcanological criteria (tectonic setting, rock geochemistry,
-volcano morphology, eruption size and eruption style), and a structured
-combination of them, to quantify overall multi-criteria (or total) volcano
-analogy among any two volcanoes in the GVP database. The data used by the
-method are extracted from the GVP database as well as from a merged database of
-volcano morphology (after Pike and Clow, 1981; Grosse et al., 2014).
 
-PyVOLCANS provides its user with full flexibility to identify customised sets
-of analogue volcanoes, by exploring three main variables:
+## Installation instructions
 
-        (1) target volcano (or volcano of interest);
-        (2) weighting scheme (i.e. set of weights given to each of the five
-        volcanological criteria to calculate multi-criteria, total analogy);
-        (3) number of 'top' analogue volcanoes (i.e. those with the highest
-        value of analogy with the target volcano).
-
-In addition, PyVOLCANS allows the user to compare the values of total analogy
-computed for 'a priori analogues' (i.e. volcanoes thought to be good analogues
-to the target volcano by other strands of evidence, e.g. expert knowledge) with
-those computed for the rest of volcanoes in the GVP database. This permits
-investigation of sets of analogue volcanoes for varied purposes, and makes
-PyVOLCANS a useful complementary method to expert-derived analogue volcanoes.
-Please see [Tierz et al. (2019)](https://doi.org/10.1007/s00445-019-1336-3) for
-more details on the VOLCANS method.
-
-## Installation
-
-PyVOLCANS can be imported to a new environment as follows:
+PyVOLCANS can be installed from PyPI as follows:
 
 ```
 pip install pyvolcans
 ```
 
-It is necessary to have Git installed and on the system path.
-
 This method adds `pyvolcans` to the virtual environment PATH so that it can be
 used from any directory.
 
 
-## How to use PyVOLCANS
+## Example usage
 
-Calling PyVOLCANS help gives details of command line options:
+Calling PyVOLCANS with a volcano name returns the 10 top analogue volcanoes:
+
+```
+$ pyvolcans Hekla
+
+Top 10 analogue volcanoes for Hekla, Iceland (372070):
+              name       country  smithsonian_id  total_analogy
+       Torfajokull       Iceland          372050       0.941676
+       Bardarbunga       Iceland          373030       0.921407
+      Prestahnukur       Iceland          371070       0.919877
+        Langjokull       Iceland          371080       0.915929
+           Hengill       Iceland          371050       0.911855
+ Brennisteinsfjoll       Iceland          371040       0.907751
+        Kverkfjoll       Iceland          373050       0.906833
+       Fremrinamar       Iceland          373070       0.905074
+           Ecuador       Ecuador          353011       0.901611
+     Marion Island  South Africa          234070       0.892960
+```
+
[…]
