### 2021-07-19 21d9f15
Small updates in README.md
With these changes, I have tried to make the development installation of the code more detailed and step-wise. I have also reworded a few sentences.
