+++ README.md
@@ -92,15 +92,22 @@ The best way to get in touch to ask questions, submit bug reports or contribute 
 
 ### For developers
 
-To modify the code, check out the repository then run the following from the root directory of
-the project to install `pyvolcans` in development mode, preferably within a clean virtual environment.
+To modify the code, first clone the PyVOLCANS repository into your local machine:
+```bash
+git clone https://github.com/BritishGeologicalSurvey/pyvolcans
+```
+
+Then move to the root directory of the project (`pyvolcans`, which contains the `setup.py` file - 
+please also check the `requirements.txt` file for a full list of dependencies in the code),
+and run the following command to install PyVOLCANS in development mode, preferably within a
+clean virtual environment:
 
 ```bash
 python -m pip install -e .[dev]
 ```
 
 The `-e` flag makes the files in the current working directory available
-throughout the virtual environment and so changes are reflected straight away.
+throughout the virtual environment and, therefore, changes are reflected straight away.
 With this installation, it is no longer required to set the PYTHONPATH.
 
 The `[dev]` part installs packages required for development e.g. `pytest`.
@@ -108,7 +115,6 @@ The `[dev]` part installs packages required for development e.g. `pytest`.
 Run tests with:
 
 ```bash
-export PYTHONPATH=.
 pytest -v test
 ```
 
