### 2021-07-20 #3436 @PTierz
Dear @elbeejay,

Very many thanks for your support and your patience. Your comments have been very useful for us to improve the documentation for our program, and to revisit the timeline and requirements needed to successfully continue with our 'translation' of the Matlab VOLCANS code, into the Python PyVOLCANS code. Many thanks.

Concerning the first point you raised, there are indeed plans of converting the entire set of Matlab scripts that calculate the volcano analogies into Python, and release them in future versions of the PyVOLCANS code. We sincerely agree with the fact that this would be a desirable addition to the program, as we acknowledge in our recently created [wiki pages](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki). However, due to the reasons explained below, we truly believe that the current submission of PyVOLCANS v1.1.2 has a substantial scientific value on its own, hence making it a suitable contribution to JOSS. Many thanks.

The timeline for the full Matlab-Python conversion might be significantly long (perhaps 1-2 years from now). This timeline would also depend on the availability of internal funding within our organisation, which should support the required team programming effort.

Nevertheless, we are always open to share further details of the VOLCANS methodology with any interested user, while we are translating the entire code into Python language, as it is now explicitly written in our [wiki pages](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki).
In a similar way, we were always completely open to the several requests we received from colleagues to compute sets of analogue volcanoes using VOLCANS, from December 2019 (when the VOLCANS work was first presented to the scientific community) to March 2021, when the first publicly available version of PyVOLCANS was released.

Since then, we have received significant [attention](https://twitter.com/volcan01010/status/1384108463505117197) and [usage](https://github.com/BritishGeologicalSurvey/pyvolcans/graphs/traffic) from a variety of potential users, which clearly supports our case to defend the relevance of the submitted PyVOLCANS v1.1.2 code. We wholeheartedly believe that making the VOLCANS results openly available to any volcano scientist and researcher in the world, holds a significant value for the scientific community (and the public). This is what the current PyVOLCANS code does, making it a working Minimum Viable Product, with reusable and testable code that enables repeatable and reproducible research. In our opinion, this stands as a very valuable first step of (ongoing) development, which we carried out, as a team, over the last year. We would also like to stress that PyVOLCANS v1.1.2 represents an enabling technology, for it allows users to carry out a great variety of scientific analyses using data-driven sets of analogue volcanoes, something that was definitely not possible before the release of the code. Very many thanks.

Concerning your second point, in order to tackle the issues you observed, we have created a series of [wiki pages](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki) inside our repository, to clearly provide every user of PyVOLCANS with detailed information on how to install, use and modify our code.

In particular, in the [README.md](https://github.com/BritishGeologicalSurvey/pyvolcans#readme) file, we have clearly described the following aspects: (1) statement of need and target audience of PyVOLCANS; (2) installation instructions, following PyPI for standard use or through cloning the GitHub repository and running our [setup.py](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/main/setup.py) file for development use of the code; (3) example usage (which is significantly expanded in the [wiki pages](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki)); (4) API documentation (we also point our future users to the source scripts and the [wiki pages](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki) to find more details on all the functions used in the code and how to run different types of analyses); (5) community guidelines, including how to contribute to the code, report issues/bugs and seek support (this information is also detailed in the [wiki pages](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki)).

We most sincerely hope that all the aforementioned additions to PyVOLCANS, together with our responses, can help our manuscript to go through the next steps in the peer-review process.

If you required anything else from us, please do not hesitate to let us know (please note that I, Pablo, will be taking summer annual leave from 24th July to 16th August. Therefore, it is unlikely that I will be able to successfully proceed with my review tasks during those weeks. Very many thanks in advance for your understanding).

Best regards,
Pablo (on behalf of the authors)
