### 2021-07-20 #3436 @elbeejay
Hi @PTierz, thank you for this explanation and thoughtful response.

I do not doubt the usefulness of the software for the community, however I still struggle slightly with the submitted software lacking the foundational computations which create the analog matrices, as well as the lack of [API documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation).

Given your leave coming up, and these outstanding doubts I have, I will be sharing this submission with the broader JOSS editorial team to get their opinions on whether or not the software and its documentation are fit for JOSS at this time, or whether the full analog matrix computations and full API documentation are needed for this submission to fit within the JOSS scope.

We will likely reach a decision by the end of next week, but given your upcoming leave, I will "pause" this pre-review and we can resume the process on August 16th.
