### 2021-07-21 bc43bca
Add docs folder and documentation markdown files
