+++ docs/source/advanced_usage.md
@@ -1,0 +1,204 @@
+# Advanced Usage
+
+## Modifying the weighting scheme and number of top analogue volcanoes
+
+Use the optional flags -Ts, -G, -M, -Sz, -St to customise the weights that PyVOLCANS assigns to each of the volcanological criteria: tectonic setting, rock geochemistry, volcano morphology, eruption size (Volcanic Explosivity Index, VEI, [Newhall and Self, 1982](https://doi.org/10.1029/JC087iC02p01231)), and eruption style, respectively (abbreviations are as in [Tierz et al., 2019](https://doi.org/10.1007/s00445-019-1336-3), Equation 1).
+
+For example, call PyVOLCANS using an equal-weight scheme between eruption size and eruption style, setting the other volcanological criteria to 0 (Scheme B in [Tierz et al., 2019](https://doi.org/10.1007/s00445-019-1336-3)):
+
+```
+$ pyvolcans Hekla -Sz 0.5 -St 0.5
+
+Top 10 analogue volcanoes for Hekla, Iceland (372070):
+               name        country  smithsonian_id  total_analogy
+              Taupo    New Zealand          241070       0.947620
+         Kikhpinych         Russia          300180       0.940538
+      Chichinautzin         Mexico          341080       0.934044
+ Carrán-Los Venados          Chile          357140       0.933013
+           Newberry  United States          322110       0.930016
+           Khodutka         Russia          300053       0.927797
+ Tangkoko-Duasudara      Indonesia          266130       0.922440
+             Dukono      Indonesia          268010       0.921541
+              Okmok  United States          311290       0.921246
+        Agua de Pau       Portugal          382090       0.921154
+```
+
+NB. You can also use fractions to define the weighting scheme. For instance, the last command above is equivalent to: `$ pyvolcans Hekla -Sz 1/2 -St 1/2`
+
+## Modifying the number of top analogues
+
+Use the optional flag --count to change the number of top analogue volcanoes:
+
+```
+$ pyvolcans Hekla --count 5
+
+Top 5 analogue volcanoes for Hekla, Iceland (372070):
+         name  country  smithsonian_id  total_analogy
+  Torfajokull  Iceland          372050       0.941676
+  Bardarbunga  Iceland          373030       0.921407
+ Prestahnukur  Iceland          371070       0.919877
+   Langjokull  Iceland          371080       0.915929
+      Hengill  Iceland          371050       0.911855
+```
+
+## Getting verbose output
+
+Use the optional flag --verbose to print the weighting scheme on-screen, and also to obtain the single-criterion analogy values, in addition to the total analogy values that PyVOLCANS outputs by default:
+
+```
+$ pyvolcans Hekla --verbose
+
+PyVOLCANS: Supplied weights: {'tectonic_setting': 0.2, 'geochemistry': 0.2, 'morphology': 0.2, 'eruption_size': 0.2, 'eruption_style': 0.2}
+Top 10 analogue volcanoes for Hekla, Iceland (372070):
+              name       country  smithsonian_id  total_analogy  ATs        AG        AM       ASz       ASt
+       Torfajokull       Iceland          372050       0.941676  0.2  0.188235  0.187584  0.180280  0.185577
+       Bardarbunga       Iceland          373030       0.921407  0.2  0.188235  0.187584  0.172727  0.172861
+      Prestahnukur       Iceland          371070       0.919877  0.2  0.188235  0.189474  0.169091  0.173077
+        Langjokull       Iceland          371080       0.915929  0.2  0.188235  0.177193  0.169091  0.181410
+           Hengill       Iceland          371050       0.911855  0.2  0.192157  0.173684  0.169091  0.176923
[…]
