+++ docs/source/contributing.md
@@ -1,0 +1,3 @@
+# Contributing to PyVOLCANS
+   
+Please feel free to contribute to the code, report any bugs or problems you encounter and/or seek support on the use and development of PyVOLCANS through the Issues tab of the repository, or through our GitHub profiles: @PTierz, @mobiuscreek, @volcan01010.
