+++ docs/source/index.rst
@@ -1,0 +1,20 @@
+Welcome to PyVOLCANS' documentation!
+=====================================
+
+`PyVOLCANS: A Python package to flexibly explore similarities and differences between volcanic systems`
+
+
+The main goal of PyVOLCANS is to help alleviate data-scarcity issues in volcanology, and contribute to developments in a range of topics, including (but not limited to): quantitative volcanic hazard assessment at local to global scales, investigation of magmatic and volcanic processes, and even teaching and scientific outreach. We hope that future users of PyVOLCANS will include any volcano scientist or enthusiast with an interest in exploring the similarities and differences between volcanic systems worldwide.
+
+Here you can find more details about PyVOLCANS, including background information, statement of need and advanced analogue searches that can be performed with the program. 
+
+Many thanks for your interest and happy analogue volcanoes searching!
+
+.. toctree::
+     :maxdepth: 2
+     :hidden:
+
+     installation
+     quickstart
+     advanced_usage
+     contributing
