### 2021-07-22 #3436 @volcan01010
Hi @elbeejay, thank you for reviewing our submission,

Since the original submission we have restructured our README.md and created the [wiki](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki) page.

The aim of PyVOLCANS is to allow relatively non-technical users the query the analogy matrix data files, so the only real user-facing API is the command-line interface.  The description of this is now more prominent on the [README.md](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/readthedocs#readme).  More complex examples are now in the wiki, where they are better-structured than they were on the single-page README.md.

The [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) say that functions/methods are documented.  While we do not have a ReadTheDocs page, or similar, they all have full Numpy-style doc strings.

Is the updated documentation suitable now?  If not, could you point to some specific examples of what you think it need, please?
