### 2021-07-22 #3436 @danielskatz
Just to interject as one of the JOSS AEiCs, the JOSS concern about API documentation is related to the fact that this is open source software where we hope that others developers will contribute to and use the software or elements of it, not just as end users.  So please don't think of our API documentation review criterion as entirely user-facing.
