### 2021-07-23 #3436 @PTierz
Hi @elbeejay,

Very many thanks for your further comments. I just would like to comment on the topics concerning the validity, usefulness, and transparency of VOLCANS and, therefore, PyVOLCANS. Many thanks in advance.

Unfortunately, it is a bit difficult for us to fully understand where our submitted manuscript fails in terms of complying with the [submission requirements]( https://joss.readthedocs.io/en/latest/submitting.html), or the scope, of JOSS. We sincerely believe that one of the greatest strengths of our submission lies on the requirement: “The software must have an **obvious** research application.” We think that this is clearly demonstrated in the [second last paragraph of our submission](https://github.com/openjournals/joss-papers/blob/joss.03436/joss.03436/10.21105.joss.03436.pdf), where up to four scientific studies over 2020-2021 have been supported by the knowledge at the base of PyVOLCANS. In early 2020, we had to make a decision between two options: either focusing on the full translation of VOLCANS from Matlab to Python, which would have required 2-3 years of development (being rather optimistic); or focusing, as an initial stage of development, on making the VOLCANS results openly available to everyone, so people could benefit from this first-of-its-kind, enabling technology to tackle issues related to data scarcity in volcanology.

If we had chosen the first route, the obvious research application of PyVOLCANS would have been kept away from the scientific community during the several years that we would have needed to develop a PyVOLCANS version that translated the entire VOLCANS method into Python. This is not to mention the enhanced risks of such a long-term project, with strong dependence on an ever-changing funding landscape, both internally in our institution, and externally, from national and/or international funding streams. In other words, chances were that the complete version of PyVOLCANS would have never been created.

Therefore, we struggle to see how the current version of PyVOLCANS is not deemed as a valuable contribution to research in volcanology, and perhaps beyond, as volcanic systems have played and play many important roles in the evolution and functioning of our geodynamic planet.

We would also like to remark that the datasets that support the creation of the analogy matrices were made fully open-access in 2019, as [Electronic Supplementary Material of the VOLCANS article]( https://link.springer.com/article/10.1007/s00445-019-1336-3#Sec25). Complementarily, the [methodology behind the calculation of the analogy matrices]( https://link.springer.com/article/10.1007/s00445-019-1336-3#Sec6) is carefully and thoroughly detailed in the aforementioned article, and it went through a comprehensive peer-review process carried out by knowledgeable scientific colleagues.

In summary, we are committed to full transparency about how PyVOLCANS derives the sets of analogue volcanoes that anyone in the world can now calculate via the Python code that we submitted for publication in JOSS. We sincerely consider that this submitted version already represents a significant step forward: i.e. a minimum viable product with obvious research applications.

Very many thanks again for your feedback and encouragement to improve our research and code development.

We will look forward to hearing from you and the editorial team of JOSS whenever it may be possible.

Receive our best regards,
Pablo (on behalf of the authors)
