### 2021-08-16 #3436 @elbeejay
Hi @PTierz, I hope you had a good holiday. I've conferred with the JOSS editorial team and am happy to say we will be moving PyVOLCANS on to the review stage.

The one recommendation we have is that the MATLAB code for generating the analog matrices should be added to the repository. This would enable an outside contributor to see the code behind that computation, or potentially perform that Matlab-Python conversion for you! Hopefully this strikes a reasonable compromise between making the available code as transparent and "open" as possible, without placing the burden of doing what sounds like an arduous conversion of code on the development team.

Can you recommend any reviewers that would be a good fit to review this work? A list of folks that have expressed a willingness to review for JOSS is available [here](https://bit.ly/joss-reviewers), but feel free to also recommend people not on this list. Please do not tag potential reviewers in this thread but rather provide their GitHub handle without the "@" symbol (e.g. "elbeejay" for me).

Thanks, let me know if you have any outstanding questions about the process.
