+++ pyvolcans/VOLCANS_matlab_scripts/README_VOLCANS.md
@@ -1,0 +1,27 @@
+# VOLCano ANalogues Search (VOLCANS)
+
+This is a README file that describes the structure and content of the Matlab scripts (Mathworks, 2012) that integrate
+the VOLCANS method presented by [Tierz, Loughlin and Calder (2019)](https://doi.org/10.1007/s00445-019-1336-3).
+
+This set of scripts was used to derive all the [analogy matrices](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/main/pyvolcans/VOLCANS_mat_files/analogy_mats)
+used by `PyVOLCANS`.
+
+Please note that the VOLCANS Matlab code is provided _as-is_: in other words, without any cleaning for public use.
+The eventual aim of us, creators of `PyVOLCANS` ([@PTierz](https://github.com/PTierz), [@mobiuscreek](https://github.com/mobiuscreek),
+[@volcan01010](https://github.com/volcan01010)), is to port the entire Matlab code into Python, including it as part of future releases
+of the `PyVOLCANS` code.
+
+In the following, a brief summary of the general process developed to generate the analogy matrices is provided (for more details,
+please see [Tierz et al., 2019](https://doi.org/10.1007/s00445-019-1336-3)). This is followed by individual descriptions of the main
+purpose and functioning of each of the Matlab (`.m`) files present in the current folder.
+
+[...]
+
+`VOLCANS` was devised by [Pablo Tierz](https://www.bgs.ac.uk/people/tierz-lopez-pablo/) ([@PTierz](https://github.com/PTierz)),
+[Susan C. Loughlin](https://www.bgs.ac.uk/people/loughlin-susan/) (British Geological Survey),
+and [Eliza S. Calder](https://www.research.ed.ac.uk/en/persons/eliza-calder) (University of Edinburgh).
+All Matlab scripts were written by [Pablo Tierz](https://www.bgs.ac.uk/people/tierz-lopez-pablo/) ([@PTierz](https://github.com/PTierz)).
+
+`PyVOLCANS` and the associated Matlab scripts are distributed under the [LGPL v3.0 licence](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/main/LICENSE).
+Copyright: © BGS / UKRI 2021.
+
