+++ pyvolcans/VOLCANS_matlab_scripts/README_VOLCANS.md
@@ -11,17 +11,137 @@ The eventual aim of us, creators of `PyVOLCANS` ([@PTierz](https://github.com/PT
 [@volcan01010](https://github.com/volcan01010)), is to port the entire Matlab code into Python, including it as part of future releases
 of the `PyVOLCANS` code.
 
-In the following, a brief summary of the general process developed to generate the analogy matrices is provided (for more details,
-please see [Tierz et al., 2019](https://doi.org/10.1007/s00445-019-1336-3)). This is followed by individual descriptions of the main
-purpose and functioning of each of the Matlab (`.m`) files present in the current folder.
+In the following, we provide: (1) a brief summary of the general process developed to generate the analogy matrices (for more details,
+please see [Tierz et al., 2019](https://doi.org/10.1007/s00445-019-1336-3)); (2) overall descriptions of the main purpose of each of the
+Matlab (`.m`) files available in the current folder; and (3) overall descriptions of the data contained in each of the csv files available
+in folder `../VOLCANS_csv_files`. Detailed specifications of the numbering adopted for categorical variables, as well as the units
+in which numerical variables are displayed (unless this is described in the header row of the csv files) are provided in
+`../VOLCANS_csv_files/VOLCANS_code_numbers`.
 
-[...]
+
+## Volcano analogy
+
+VOLCANS interprets volcano analogy as a quantitative measure of the similarity between any two volcanic systems listed in the Holocene
+[Volcanoes Of The World database](https://volcano.si.edu/list_volcano_holocene.cfm) (v. 4.6.7), hosted by the Global Volcanism Program
+(GVP) of the Smithsonian Institution. A volcano analogy of 0 denotes _no analogy_ between the two volcanoes, while a volcano analogy of
+1 implies that the volcanoes are _perfect analogues_, considering the data available.
+
+Five different volcanological criteria are considered in VOLCANS: tectonic setting, rock geochemistry, volcano morphology, eruption size
+and eruption style. Volcano analogies based on only one criterion are termed: ''single-criterion analogies''. Volcano analogies based on
+a structured combination (a weighted average) of two or more volcanological criteria are termed: ''multi-criteria or total analogies''.
+Provided that each single-criterion analogy is a number between 0-1, and that the sum of weights defining the weighted average is equal
+to 1, then each multi-criteria, or total analogy, is also a number between 0-1 (please see [Methods](https://link.springer.com/article/10.1007/s00445-019-1336-3#Sec6)
+in Tierz et al., 2019, for more details). Please note that if there is no data available for a given volcano and a particular volcanological
+criterion, then the corresponding single-criterion analogy values between that volcano and any other volcano in the GVP database are set to 0.
+
+VOLCANS uses distance metrics to calculate each set of single-criterion analogies between any two volcanic systems in the GVP database.
+These distance metrics are either based on: (a) linear distances between volcano-specific values, when the criterion is described by a
+variable with single values for each volcano (tectonic setting, volcano morphology); (b) areal differences between volcano-specific Empirical
+Cumulative Distribution Functions (ECDFs), when the criterion is described by a variable with a probability distribution (or histogram) for
+each volcano (rock geochemistry, eruption size); or (c) normalised sum of differences between the frequency of occurrence of different groups
[…]
