### 2021-08-20 34baa49
Add csv data files and documentation on their code numbering
In this commit, we are incorporating all data sources (csv files) used by VOLCANS to derive
the analogy matrices that are then handled by PyVOLCANS. We also include documentation on the
code numbering adopted in the csv files, which is then used to perform the calculations in the
VOLCANS Matlab scripts.
