+++ pyvolcans/VOLCANS_csv_files/VOLCANS_code_numbers/README_NUMBERING.md
@@ -1,0 +1,113 @@
+# VOLCANS: code numbering
+
+This README file details the code numbering adopted in the different csv files inside the folder:
+`../../VOLCANS_csv_files`. This numbering was adopted to ease the calculations performed by the Matlab
+scripts that compose the VOLCANS method.
+
+Please note that, in some cases, the code numbering can be directly checked on `.xls` files located inside
+the current folder (please see descriptions below). When this is not available, a legend is provided in this
+file, showing the correspondence between categorical variables and code numbering (please see next sub-sections).
+
+Please also note that the `NO DATA` value used across all data files is `-9999`.
+
+## VOTW467_8May18_volcano_data.csv
+
+Code numbering for all the categorical variables can be found in the file `VOTW467_8May18_Holocene_list_textdata_portable.xls`
+(the variable displaying the code number in the spreadsheet is given between brackets):
+
+Country (country#), Region (reg#), Subregion (subreg#), Primary Volcano Type (volctype#), Activity Evidence (procdate1),
+Dominant Rock Type (domrock#), Tectonic Setting (tectset#).
+
+Please also note the following four aspects:
+
+1. The variable ''Volcano Number'' (also expressed as ''VNUM'' in other data files) denotes a unique volcano identifier that
+the GVP database assigns to each volcanic system listed in the database.
+
+2. The variables ''Major Rock 1-5'' and ''Minor Rock 1-5'' use the same code numbering as the variable ''Dominant Rock Type''.
+
+3. The variable ''Last Known Eruption'' is given in years from current era, with negative values expressing dates before
+current era (BCE). The same convention is applicable to variables ''Start Year'' in file `VOTW467_8May18_eruption_data.csv`,
+''Eruption Start Year'' in file `VOTW467_8May18_event_data.csv`, and ''Date of completeness'' in `MeadMagill2014_June2018.csv`.
+
+4. The variables ''Latitude'' and ''Longitude'' (also included in the file `VOTW467_8May18_Eruption_list_textdata_portable.xls`)
+correspond to the approximate spatial location of the volcanic system, and are expressed in decimal degrees.
+
+## VOTW467_8May18_eruption_data.csv
+
+Code numbering for the categorical variable ''VEI Modifier'' can be found in the file `VOTW467_8May18_Eruption_list_textdata_portable.xls`
+(please see column ''VEImod#'').
+
+For the variable ''Eruption Category'', the following code numbering is applied:
+
+1 = Confirmed Eruption
+
+3 = Uncertain Eruption 
+
+## VOTW467_8May18_event_data.csv
+
+Code numbering for the categorical variable ''Event Type'' can be found in the file `VOTW467_8May18_Event_list_textdata_portable.xls`
+(please see column ''event#'').
+
+For the variable ''group#'', the following code numbering is applied (please also see [Tierz et al., 2019](https://doi.org/10.1007/s00445-019-1336-3)):
+
+1 = Lava flow and/or fountaining
+
+2 = Ballistics and Tephra
+
+3 = Phreatic and phreatomagmatic activity 
+
+4 = Water-sediment flows
+
+5 = Tsunamis
+
+6 = Pyroclastic density currents
+
+7 = Edifice collapse/destruction
+
+8 = Caldera formation
+
+## MeadMagill2014_June2018.csv
+
+Code numbering for the categorical variable ''Region/Country code'' corresponds with the values in column ''Name''.
+These values are the same as those found in file `VOTW467_8May18_Holocene_list_textdata_portable.xls`, for variables
[…]
