+++ pyvolcans/VOLCANS_matlab_scripts/eruption_style_processing.m
@@ -1,0 +1,192 @@
+% eruption_style_processing.m
+
+%{Summary:
+
+It derives the data required to calculate the single-criterion analogy matrix for eruption style.
+It requires to import the data in the files: `VOTW467_8May18_event_data.csv`,
+`VOTW467_8May18_eruption_data.csv`, and `VOTW467_8May18_volcano_data.csv` to
+carry out the aforementioned task.
+
+%}
+
+clear all
+close all
+
+%load the events data
+a=importdata('../VOLCANS_csv_files/VOTW467_8May18_event_data.csv');
+
+%we load the eruption data as well, to exclude caldera formation linked
+%with VEI<4 eruptions
+b=importdata('../VOLCANS_csv_files/VOTW467_8May18_eruption_data.csv');
+
+%import the volcano data as well, we will need to work with it too
+c=importdata('../VOLCANS_csv_files/VOTW467_8May18_volcano_data.csv');
+
+%here we create the matrices that will store the main-group counts and the
+%normalised values (by eruption). Note that we will not need any type of
+%histogram this time
+AStcountmat=zeros(length(c.data(:,1)),8);  %there are 8 main groups
+AStnormmat=zeros(length(c.data(:,1)),8);
+
+%here we include the analysis of phreatomagmatic eruptions only
+phreatomag_norm=zeros(length(c.data(:,1)),1);
+
+%we create the loop that we will use to populate the matrices
+for ii=1:length(AStcountmat(:,1))
+    vnum=c.data(ii,1);  %getting the VNUM of the volcano
+        
+    idsvolc=find(a.data(:,1)==vnum); %finding all entries (events) that are
+                                     %linked with the specific volcano
+    
+    %finding the total number of unique eruptions from all the events
+    %(i.e. length(ia_er))
+    [Cer,ia_er,ic_er]=unique(a.data(idsvolc,3),'first');
+    
+    Ner=length(ia_er); %total number of eruptions
+    
+    %%%fprintf('VNUM=%d; Ner=%d\n',vnum,Ner);
+    
+    %explore the eruptions searching for the main groups of events
+    for jj=1:Ner
+        ernum=a.data(idsvolc(ia_er(jj)),3); %we deal with 1 eruption at a time
+        idserupt=find(a.data(:,3)==ernum); %global indices for the eruption
+        
+        %get the counts (max. 1) per each main group and eruption (NB. We
+        %must use find because knnsearch gives the nearest neighbour, hence
+        %it will never return an empty matrix!
+        lava=find(a.data(idserupt,7)==1);
+        tephra=find(a.data(idserupt,7)==2);
+        phreato=find(a.data(idserupt,7)==3);
+        wsedflow=find(a.data(idserupt,7)==4);
+        tsunami=find(a.data(idserupt,7)==5);
+        pdc=find(a.data(idserupt,7)==6);
+        edestru=find(a.data(idserupt,7)==7);
+        caldera=find(a.data(idserupt,7)==8);
+        
+        %here we deal with the phreatomagmatic eruptions only
+        phreatomag_only=find(a.data(idserupt,6)==17);
+    
+        %get the counts (if the vectors are not empty)
+        if ~isempty(lava)
+            AStcountmat(ii,1)=AStcountmat(ii,1)+1;
+        end
+        
+        if ~isempty(tephra)
+            AStcountmat(ii,2)=AStcountmat(ii,2)+1;
+        end
+        
+        if ~isempty(phreato)
[…]
