+++ pyvolcans/VOLCANS_matlab_scripts/votw_analysis.m
@@ -1,0 +1,602 @@
+% votw_analysis.m
+
+%{Summary:
+
+It performs a preliminary analysis of the volcano data in the GVP database (v4.6.7),
+and derives the data required to calculate the single-criterion analogy matrices for
+tectonic setting and rock geochemistry. It requires to import the data in the file
+`VOTW467_8May18_volcano_data.csv` to carry out the aforementioned tasks.
+
+%}
+
+clear all
+close all
+
+%importing the data
+a=importdata('../VOLCANS_csv_files/VOTW467_8May18_volcano_data.csv');
+
+%coarse separation of tectonic settings
+intpt=find(a.data(:,13)==1 | a.data(:,13)==2 | a.data(:,13)==3);
+rift=find(a.data(:,13)==4 | a.data(:,13)==5 | a.data(:,13)==6);
+subd=find(a.data(:,13)==7 | a.data(:,13)==8 | a.data(:,13)==9 | ...
+    a.data(:,13)==10);
+
+alltect=[intpt;rift;subd];  %concatenating tectonic settings
+
+%plotting volcanoes by tectonic setting
+load coast
+figure(); plot(long,lat,'k');
+hold on; scatter(a.data(:,10),a.data(:,9),'r^','filled');
+hold on; scatter(a.data(rift,10),a.data(rift,9),'g^','filled');
+hold on; scatter(a.data(subd,10),a.data(subd,9),'b^','filled');
+legend({'coastline','intraplate','rift','subduction'},'FontSize',10,...
+    'location','best');
+set(gca,'FontSize',15,'LineWidth',1.5);
+set(gcf,'PaperSize',[20 16],'PaperUnits','centimeters',...
+'PaperPosition',[1 1 16 12],'PaperOrientation','portrait');
+box on;
+xlabel('Longitude [deg]','fontsize',14);
+ylabel('Latitude [deg]','fontsize',14);
+print -f1 -dpng -r300 volcanoes_by_tectosett.png
+
+%histogram of simple tectonic types: intrapt, rift, subduction
+figure(); hist(a.data([intpt;rift;subd],13),[2 5 8.5]);
+
+%plotting volcanoes by primary rock type
+ande=find(a.data(:,12)==1);
+bas=find(a.data(:,12)==2);
+dac=find(a.data(:,12)==3);
+rhy=find(a.data(:,12)==8);
+alk=find(a.data(:,12)==4 | a.data(:,12)==6 | a.data(:,12)==7 | ...
+    a.data(:,12)==9 | a.data(:,12)==10 | a.data(:,12)==11);
+
+allrock=[bas;ande;dac;rhy;alk]; %concatenating all (dominant) rock types
+
+%let's deal with all rock types here as well
+%we start with finding out the volcanoes that lack each major and minor
+%rock types (max. 5 types per category)
+noMR1=find(a.data(:,14)==-9999);    nomR1=find(a.data(:,19)==-9999);
+noMR2=find(a.data(:,15)==-9999);    nomR2=find(a.data(:,20)==-9999);
+noMR3=find(a.data(:,16)==-9999);    nomR3=find(a.data(:,21)==-9999);
+noMR4=find(a.data(:,17)==-9999);    nomR4=find(a.data(:,22)==-9999);
+noMR5=find(a.data(:,18)==-9999);    nomR5=find(a.data(:,23)==-9999);
+
+
+figure(); plot(long,lat,'k');
+hold on; scatter(a.data(bas,10),a.data(bas,9),'k^','filled');
+hold on; scatter(a.data(ande,10),a.data(ande,9),'b^','filled');
[…]

[…]
