+++ pyvolcans/VOLCANS_matlab_scripts/README_VOLCANS.md
@@ -13,10 +13,10 @@ of the `PyVOLCANS` code.
 
 In the following, we provide: (1) a brief summary of the general process developed to generate the analogy matrices (for more details,
 please see [Tierz et al., 2019](https://doi.org/10.1007/s00445-019-1336-3)); (2) overall descriptions of the main purpose of each of the
-Matlab (`.m`) files available in the current folder; and (3) overall descriptions of the data contained in each of the csv files available
-in folder `../VOLCANS_csv_files`. Detailed specifications of the numbering adopted for categorical variables, as well as the units
+Matlab (`.m`) files available in the [current folder](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/matlab-scripts/pyvolcans/VOLCANS_matlab_scripts); and (3) overall descriptions of the data contained in each of the csv files available
+in folder [../VOLCANS_csv_files](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/matlab-scripts/pyvolcans/VOLCANS_csv_files). Detailed specifications of the numbering adopted for categorical variables, as well as the units
 in which numerical variables are displayed (unless this is described in the header row of the csv files) are provided in
-`../VOLCANS_csv_files/VOLCANS_code_numbers`.
+[../VOLCANS_csv_files/VOLCANS_code_numbers](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/matlab-scripts/pyvolcans/VOLCANS_csv_files/VOLCANS_code_numbers).
 
 
 ## Volcano analogy
@@ -44,7 +44,7 @@ and [Table 2](https://link.springer.com/article/10.1007/s00445-019-1336-3/tables
 
 ## Matlab scripts
 
-The following Matlab scripts have the purpose of calculating volcano analogy matrices that contain the single-criterion volcano analogy values
+The following [Matlab scripts](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/matlab-scripts/pyvolcans/VOLCANS_matlab_scripts) have the purpose of calculating volcano analogy matrices that contain the single-criterion volcano analogy values
 for each of the possible combinations of two volcanoes in the GVP database. All the analogy matrices are N x N, where N (= 1439) is the total number of
 volcanic systems listed in the GVP v4.6.7 database used in [Tierz et al. (2019)](https://doi.org/10.1007/s00445-019-1336-3).
 Please also note that any single-criterion analogy between volcanoes X and Y, e.g. rock geochemistry (AG<sub>XY</sub>), is symmetric.
@@ -52,54 +52,52 @@ That is: AG<sub>XY</sub> = AG<sub>YX</sub>.
 
 ### get_final_AT_allcross.m
 
-It calculates the final single-criterion analogy matrix for tectonic setting. It requires to load the file `ATmatrices.mat`, which is derived
-from the script `votw_analysis.m`.
+It calculates the final single-criterion analogy matrix for tectonic setting. It requires to load the file [ATmatrices.mat](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/matlab-scripts/pyvolcans/VOLCANS_mat_files/data_mats/ATmatrices.mat), which is derived
+from the script [votw_analysis.m](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/matlab-scripts/pyvolcans/VOLCANS_matlab_scripts/votw_analysis.m).
 
 ### get_final_AG_allcross.m
 
-It calculates the final single-criterion analogy matrix for rock geochemistry. It requires to load the file `AGmatrices_ALU_QUET.mat`, which is
-derived from the script `votw_analysis.m`.
[…]
