### 2021-08-20 #3436 @PTierz
Hi @elbeejay,

Many thanks for your message and your wishes (holidays were great, thanks). Apologies for the delay in replying.

It is very positive news to hear that our submission of `PyVOLCANS` can continue into the review stage. I thank you and the JOSS editorial team for your work and support.

We are happy to comply with your recommendation and upload the Matlab scripts that I used to generate the analogy matrices. Your point about outside contributors being able to see the Matlab code and collaborate to its conversion into Python is an important one. Many thanks for your feedback. We sincerely hope that, through collaboration with the scientific community, `PyVOLCANS` can become an increasingly useful tool for volcanologists and other scientists.

I am glad to inform you that I have just uploaded the relevant [set of Matlab scripts](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/matlab-scripts/pyvolcans/VOLCANS_matlab_scripts), together with a [README_VOLCANS.md file](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/matlab-scripts/pyvolcans/VOLCANS_matlab_scripts/README_VOLCANS.md) that documents some extra details on `VOLCANS` and on the concept of _volcano analogy_, and provides a brief description of the main purpose of each of the scripts uploaded. It also contains details of what data are stored in each of the csv files that are used by the scripts (please find additional descriptions of the data contained in the csv files in the [README_NUMBERING.md file](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/matlab-scripts/pyvolcans/VOLCANS_csv_files/VOLCANS_code_numbers/README_NUMBERING.md) that we have also created. Many thanks).

We most sincerely hope that these additions can be satisfactory. If you required anything else from us, please do not hesitate to let us know. Many thanks.

Finally, in terms of reviewers, we consider that the following three could be a good fit for our manuscript: jifarquharson, weiji14, rbeucher.

I believe that we do not have any specific questions at this time, but many thanks for your availability.

Receive our best regards,

Pablo (on behalf of the authors)
