### 2021-08-22 #3436 @elbeejay
@PTierz great! I think this works well and gives the interested user/reader an opportunity to both see the underlying code and contribute to the overall project if they so desire. Thank you for being so accommodating.
