### 2021-08-22 #3436 @elbeejay
:wave: @jifarquharson, @weiji14, @rbeucher, would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

The submission is "PyVOLCANS: A Python package to flexibly explore similarities and differences between volcanic systems". I am reaching out to you as you all are geoscientists with experience coding and working in Python.

This is a pre-review issue which is used to find reviewers. Once 2 or 3 reviewers have been found I'll start the review on a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review but can recommend someone else, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested in reviewing, please take a look at the journal's [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) policy to ensure you do not have a conflict before agreeing to review this submission.

Thank you, feel free to reach out if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time.

-Jay
