### 2021-08-23 #3436 @rbeucher
Note that I don't have much expertise on volcanoes. So if @meghanrjones is happy to review...
