### 2021-08-23 #3436 @meghanrjones
@elbeejay, I could review this submission. I do not have prior experience with JOSS but have expertise in OSS and volcanology.
