### 2021-08-23 #3649 @elbeejay
Hi @jifarquharson and @meghanrjones, thank you both again for agreeing to review this submission! We are currently asking reviewers to try and complete their reviews in 6 weeks. whedon will automatically send you a reminder after 2 weeks, and I'll be checking in throughout the process to answer any questions. If you find you need more time to review just say so in the thread here.

The JOSS review process is entirely open and transparent, and takes place on GitHub. Review comments should be made as issues in the PyVOLCANS repository, please just link this review issue when doing so to help us keep track of things (paste `https://github.com/openjournals/joss-reviews/issues/3649` into the issue text).

For reference, links to the JOSS reviewer criteria and checklist are below as well:
- [JOSS Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)
- [JOSS Review Checklist](https://joss.readthedocs.io/en/latest/review_checklist.html)

Feel free to ping me (@elbeejay) if you have any questions or concerns. Thanks again for agreeing to review for JOSS.
