### 2021-09-03 2ee5026
Update MANIFEST.in
The changes in this commit have the purpose of including the VOLCANS Matlab scripts and the source data csv files in the PyVOLCANS installation.
