### 2021-09-03 b771940
Fix update on MANIFEST.in
