### 2021-09-16 BritishGeologicalSurvey/pyvolcans#11 @jifarquharson
Volcano naming conventions
### Summary ###
`pyVOLCANS` documentation could do with a note on GVP naming conventions.

### Description ###
Many volcanoes have names that are more than one word, e.g. Rincon de la Vieja (Costa Rica) or St. Helens (USA): as it stands, if a user were to enter `$ pyvolcans Rincon de la Vieja` (consistent with the examples given, i.e. `$ pyvolcans Hekla`), `pyvolcans` will throw an `"unrecognized argument"` error. This can be avoided simply by wrapping the multi-word string in quotes, but as far as I can tell this isn't mentioned in the documentation.

On a related note, volcano names crawled from the Smithsonian's GVP are often in a {Name, Prefix} format, which might not be immediately obvious for a lay end-user. A common example is "Fournaise, Piton de la." It seems that the fuzzy matching does a reasonable job of catching these in most cases; for example, "Fournaise, Piton de la" is the 9th alternative option given by `fuzzy_matching()` for "Piton de la Fournaise," and "Sawad, Harra Es-" is the 4th option given for "Harra Es-Sawad," and so on. Other systems don't fare as well (e.g. Cerro Bravo or Jabal Yar), and don't offer the correct option. While arbitrarily increasing the fuzzy limit, e.g. by including
`limit = 10 if len(volcano_name.split()) == 1 else 20`
inside the `fuzzy_matching` function helps to a degree, this workaround still doesn't capture a dozen or so systems (for example, "Bravo, Cerro" is the 42nd-best fuzzy match for "Cerro Bravo"). While writing a targeted workaround for this is probably overkill given that it is potentially a niche problem, it would be useful to have some note in the documentation regarding how to input names, the format they should be in, and where to look if your volcano appears not to be on the list, e.g. due to formatting, punctuation, or usage of a different name (one can imagine researchers looking for "Fagradalsfjall," for example, rather than "Krýsuvík-Trölladyngja").

### Acceptance criteria ###

- [x] Updated documentation/instructions

Ping [3649](https://github.com/openjournals/joss-reviews/issues/3649)
