### 2021-09-21 66305f1
Change default fuzzy matching function to token_sort_ratio
