### 2021-09-22 aae8462
Update in README.md (volcano naming)
The additional explanations now included in the [README.md file](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/main/README.md) have the goal of clarifying some aspects related to the volcano naming used by GVP, and adopted in `PyVOLCANS`. This follows the recommendations provided by @jifarquharson in Issue #11 (many thanks again for your feedback).
