### 2021-09-22 BritishGeologicalSurvey/pyvolcans#11 @PTierz
Dear @jifarquharson,

Thank you very much for your comments and suggestions. We have carefully analysed them and, hopefully, fixed them to a satisfactory level.

Let us start with the issue observed with the `fuzzy_matching()` function. We have realised that just by substituting the function used to calculate the `scorer` value in the call to `process.extract()`, we were able to fix the issues related to the ordering of words within a given `volcano_name`. Please see commit 66305f1 (many thanks @mobiuscreek for the fix!). Therefore, in the current version of PyVOLCANS ([v1.2.2]( https://github.com/BritishGeologicalSurvey/pyvolcans/releases/tag/v1.2.2)), all the described searches provide the appropriate `volcano_name` as the first entry of the `fuzzy_matching()` output: `$ pyvolcans “Piton de la Fournaise”`, `$ pyvolcans “Harra Es-Sawad”`, `$ pyvolcans “Jabal Yar”`, and `$ pyvolcans “Cerro Bravo”`. The last example is a complicated one, as 17 volcanoes in the latest [Holocene Volcano List]( https://volcano.si.edu/list_volcano_holocene.cfm) include the word “Cerro” in their volcano names. We have performed further checks using: `$ pyvolcans "Cerro Tujle"`, `$ pyvolcans "Cerro del Azufre"`, and `$ pyvolcans "Cerro el Ciguatepe"`, and, in all cases, the first returned suggestion corresponds with the `volcano_name` of the target volcano. Many thanks.

Regarding the use of quotation marks for any `volcano_name` that is composed of more than one word, very many thanks for pointing it out – we had not thought of including these instructions. We have now included a few sentences in the main [README.md file]( https://github.com/BritishGeologicalSurvey/pyvolcans/blob/main/README.md) that explain this matter and how to properly run `pyvolcans` in this kind of situations.

Similarly, we have included some explanations about possible synonyms and subfeatures linked with the volcanic systems listed in the GVP database, and have suggested a possible workaround to this problem.

We sincerely hope that the improvements incorporated in the latest version of PyVOLCANS will be satisfactory, and this can close the current open issue. Very many thanks.

In any case, I will leave the acceptance criteria unchecked, so you can have the chance to double-check what I have described above. Could I kindly ask you (after your checks) to check the box and close the issue if you consider that our changes have resolved the issue?

Thank you very much again for your reviews and suggestions.

Best regards,
Pablo (on behalf of the co-authors)
