### 2021-09-22 BritishGeologicalSurvey/pyvolcans#11 @jifarquharson
Hi @PTierz,

Thanks for digging into this. I gave [v1.2.2](https://github.com/BritishGeologicalSurvey/pyvolcans/releases/tag/v1.2.2) a test-drive, and the one-line fix at [`66305f1`](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/66305f13dc8b457fadc463e25aea43b26411801f) works great! Combined with the updates to the README, the peculiarities of the naming convention are now much easier to navigate. Happy to close this issue.

Best Regards,
Jamie
