### 2021-09-23 c5adc73
Remove returned variable my_volcano_data_dictionary from calculate_weighted_analogy_matrix()
This change implied further changes on pyvolcans.py and
test_pyvolcans.py. NB. The `check_for_criteria_without_data()`
function does not need to be imported in pyvolcans.py anymore.
