### 2021-09-23 9398f8b
Substitute PyvolcansError by warning.warn() for non-data criteria
