### 2021-09-23 ebdb471
Substitute PyvolcansError by warning.warn() for 'perfect analogues'
