### 2021-09-23 c567832
Improve formatting of warning messages
A new, very simple function (`format_warning_message()`) has been
created to handle the formatting of the warning messages, instead of
using `warnings.warn()`. In the documentation of the functions that
check for missing-data volcanological criteria and the occurrence of
'perfect analogues', outdated references to the raise of PyvolcansError
have been deleted.
