### 2021-09-29 #3649 @meghanrjones
> Hello all, looks like @jifarquharson has been opening up some issues and has finished (or nearly finished) their review of this submission. @meghanrjones I wanted to check in to see how things are looking on your end. When do you anticipate being able to review PyVOLCANS?
>
> Thanks all and thanks to @PTierz and co-authors for both their submission and the revisions they have made thus far in response to review comments.

Hi @elbeejay, thanks for checking in. I'll provide a review within the next week.
