### 2021-09-30 96e97c3
Update paper.md to correct typo
The only change that is included in the revised version is the correction of a typo in the "Acknowledgements" section, where the name of Prof. Einat Lev had been misspelled. Many thanks @jifarquharson for spotting the mistake!
