### 2021-09-30 #3649 @PTierz
Dear @elbeejay, @jifarquharson and @meghanrjones,

Thank you very much for your messages and for all your work on the review of our manuscript.

@jifarquharson - We are hugely thankful for your very positive comments on `PyVOLCANS` and its potential applications. Regarding the typo, many thanks for pointing it out. I have corrected it in the [paper.md](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/JOSS_paper/paper/paper.md) file, in the `JOSS_paper` branch (changes are highlighted in bold text). Finally, concerning [Issue #4](https://github.com/BritishGeologicalSurvey/pyvolcans/issues/4), we hope we will have closed it by the end of this week. Please do not hesitate to let us know if you require anything else from us.

Very many thanks all once again.

Best regards,
Pablo (on behalf of the co-authors)
