### 2021-09-30 4f992dc
Simplify search for criteria without data
