### 2021-10-01 166619a
Improve warning message by adding input flag list
