### 2021-10-05 3f28ef6
Use Python UserWarning for warnings
Replacing the formatwarning function allows us to define the warning
style while still raising genuine warnings.  This is useful because
users may want to filter/suppress the messages as we have done with the
FuzzyWuzzy messages.

Renaming the "check_for..." functions to "warn_on..." makes their
purpose more explicit.
