+++ pyvolcans/pyvolcans_func.py
@@ -24,6 +24,19 @@ from pyvolcans import (load_tectonic_analogy,
                        load_eruption_style_analogy,
                        load_volcano_names)
 
+
+# Define custom message formatter for warnings
+def _format_pyvolcans_warning(message, *args, **kwargs):
+    """
+    Format warnings into PyVOLCANS style, discarding all but the
+    message.
+    """
+    return f"UserWarning (PyVOLCANS): {message}\n"
+
+
+# Replace built-in warning format function
+warnings.formatwarning = _format_pyvolcans_warning
+
 # fuzzywuzzy would like to use a sequence matcher provided by the
 # Python-Levenshtein package, but this has dependencies that require
 # compilation.  When it is not installed, it uses the matcher provided
@@ -385,9 +398,9 @@ def calculate_weighted_analogy_matrix(my_volcano, weights,
                 my_volcano_single_analogies[volcano_idx]
 
     # check for volcanological criteria without data for target volcano
-    check_for_criteria_without_data(my_volcano_data_dictionary,
-                                    my_volcano,
-                                    weights)
+    warn_on_criteria_without_data(my_volcano_data_dictionary,
+                                  my_volcano,
+                                  weights)
 
     # calculate single-criterion analogy matrices for specific weighting scheme
     weighted_tectonic_analogy = \
@@ -480,27 +493,11 @@ def get_analogies(my_volcano, volcans_result, count=10):
 
     return filtered_result, volcano_name
 
-def format_warning_message(message):
-    """
-    Takes a warning message and prints it in an appropriately formatted way.
 
-    Parameters
-    ----------
-    message: str
-        Warning message to be displayed after formatting.
-    """
-
-    formatted_warning = \
-            warnings.formatwarning(message, UserWarning,
-                                   filename=None, lineno=None)
-    formatted_warning_split = formatted_warning.split(':')
-    print(f'PyVOLCANS: {formatted_warning_split[-1]}')
-
-
-def check_for_perfect_analogues(result):
+def warn_on_perfect_analogues(result):
     """
     Assesses whether all the calculated top analogue volcanoes share the same
-    value of total analogy, and raises a PyvolcansError exception if that is
+    value of total analogy, and raises UserWarning if that is
     the case.
 
     Parameters
@@ -513,20 +510,19 @@ def check_for_perfect_analogues(result):
 
     maximum_analogy = result['total_analogy'].iloc[0]
     if result['total_analogy'].eq(maximum_analogy).all():
-        msg = ("WARNING!!! "
-               "All top analogue volcanoes have the same value "
+        msg = ("All top analogue volcanoes have the same value "
                "of total analogy. Please be aware of possible "
                "data deficiencies and/or the use of a simplified "
                "weighting scheme (see Tierz et al., 2019, for more "
                "details).\n")
-        format_warning_message(msg)
+        warnings.warn(msg)
 
 
-def check_for_criteria_without_data(my_volcano_data, my_volcano_name, weights):
+def warn_on_criteria_without_data(my_volcano_data, my_volcano_name, weights):
     """
     Assesses whether some volcanological criteria do not have any data for the
-    specific target volcano chosen by the user, raising a PyvolcansError
-    exception if this is the case, informing the user which are these criteria.
+    specific target volcano chosen by the user, raising a UserWarning
+    if this is the case, informing the user which are these criteria.
 
     Parameters
     ----------
@@ -540,7 +536,7 @@ def check_for_criteria_without_data(my_volcano_data, my_volcano_name, weights):
     weights : dict
         Set of weights (weighting scheme) selected by the user to run PyVOLCANS
     """
[…]
