### 2021-10-05 03a809a
Flake8 fixes
