### 2021-10-05 cbe91cd
Add test for warnings on missing criteria
