### 2021-10-06 02b83cb
Minor modifications in warning text formatting
