### 2021-10-11 BritishGeologicalSurvey/pyvolcans#16 @meghanrjones
Add option to report the metrics for the target and analog volcanoes
This feature request is to allow the user to query the variable values for each category for the target and/or analog volcanoes in addition to the analogy metrics. For example, a flag that would report `Ts: 0.1250; Ip - Oceanic crust` etc. for Kilauea).
