### 2021-10-13 b5dffb6
Add DOI and change publication year for Simmons (2021)
