+++ paper/paper.bib
@@ -71,11 +71,12 @@ school={University of Edinburgh}
   url={https://doi.org/10.3389/feart.2014.00028}
 }
 
-@phdthesis{Simmons:2020,
+@phdthesis{Simmons:2021,
   title={The {Quetrupillán Volcanic Complex, Chile: Holocene} volcanism, magmatic plumbing system, and future hazards},
   author={Isla Simmons},
-  year={2020},
-  school={University of Edinburgh}
+  year={2021},
+  school={University of Edinburgh},
+  url={https://doi.org/10.7488/era/1087}
 }
 
 @techreport{Crummy:2021,
