+++ paper/paper.md
@@ -64,7 +64,7 @@ ATs: Analogy in Tectonic setting; AG: Analogy in rock Geochemistry; AM: Analogy 
 \*Number between brackets denotes the unique volcano identifier used by the GVP database.
 \label{fig:figure1}](figure.png) 
 
-The results from the VOLCANS study have already been used in recent research: e.g. exploring the volcanological factors that influence the development of particular volcano morphologies [@White:2020]; constraining potential hazardous phenomena and hazard scenarios at a given target volcano, based on its analogue volcanoes [@Simmons:2020]; quantifying probability distributions of eruption sizes and probabilities of occurrence of diverse hazardous phenomena [@Tierz:2020]; or even exploring volcano analogies at regional scales, by generating sets of analogue volcanoes for tens of volcanic systems [@Crummy:2021].
+The results from the VOLCANS study have already been used in recent research: e.g. exploring the volcanological factors that influence the development of particular volcano morphologies [@White:2020]; constraining potential hazardous phenomena and hazard scenarios at a given target volcano, based on its analogue volcanoes [@Simmons:2021]; quantifying probability distributions of eruption sizes and probabilities of occurrence of diverse hazardous phenomena [@Tierz:2020]; or even exploring volcano analogies at regional scales, by generating sets of analogue volcanoes for tens of volcanic systems [@Crummy:2021].
 The last two example applications have played a key role in developing quantitative hazard analyses for Ethiopian volcanoes, within the [RiftVolc project](https://www.bgs.ac.uk/geology-projects/volcanoes/riftvolc/).
 
 We hope that the release of `PyVOLCANS` will encourage studies based on data-driven selection of analogue volcanoes and that such analyses will continue to grow in number and diversity of their scientific purposes.
