### 2021-10-14 6e4d366
Load and preliminarily check volcano data
In this commit, the .mat files containing all the volcano-specific
data used to calculate all the single-criterion analogies for the
~1400 volcanoes in GVP v4.6.7 are loaded through `base.py`, and
subsequently, the volcano "ID profile" for the selected target
volcano is printed inside the `calculate_weighted_analogy_matrix()`
function (this is just a preliminary test of intended outputs. The
initial idea is to create a dedicated function that can print this
"ID profile" to the stdout, and maybe also produce a plot showing
these data in a more clear for
[…]
