+++ pyvolcans/VOLCANS_mat_files/base.py
@@ -12,6 +12,7 @@ from pymatreader import read_mat
 BASE_DIR = Path(__file__).resolve().parent
 DIRNAME_ANALOGY = BASE_DIR.joinpath("analogy_mats")
 DIRNAME_VOLCANO = BASE_DIR.joinpath("VOTW_prepared_data")
+DIRNAME_DATA = BASE_DIR.joinpath("data_mats")
 
 
 def load_tectonic_analogy():
@@ -36,3 +37,18 @@ def load_eruption_style_analogy():
 
 def load_volcano_names():
     return pd.read_csv(DIRNAME_VOLCANO.joinpath("volc_names.csv"), header=None)
+
+def load_tectonic_data():
+    return read_mat(DIRNAME_DATA.joinpath("ATmatrices.mat"))["ATlast"]
+
+def load_geochemistry_data():
+    return read_mat(DIRNAME_DATA.joinpath("AGmatrices_ALU_QUET.mat"))["AGnormmat"]
+
+def load_morphology_data():
+    return read_mat(DIRNAME_DATA.joinpath("AMmatrices_QUET.mat"))["AMlast"]
+
+def load_eruption_size_data():
+    return read_mat(DIRNAME_DATA.joinpath("ASzmatrices_SINA.mat"))["ASznorm_ALL"]
+
+def load_eruption_style_data():
+    return read_mat(DIRNAME_DATA.joinpath("AStmatrices_SINA.mat"))["AStnormmat"]
