### 2021-10-15 f2c5203
Remove Crummy et al. (2021) from bibliography
