+++ paper/paper.bib
@@ -79,15 +79,6 @@ school={University of Edinburgh}
   url={https://doi.org/10.7488/era/1087}
 }
 
-@techreport{Crummy:2021,
-  title={Estimation of the frequency and magnitude of potential future eruptions from {Ethiopian} volcanoes through expert elicitation},
-  author={Crummy, JM and Vye-Brown, C and Marchant, B and Leeming, K and Loughlin, SC},
-  year={2021},
-  institution={British Geological Survey Internal Report},
-  volume={OR/21/013},
-  pages={43}
-}
-
 @book{Siebert:2010,
   title={{Volcanoes of the World}},
   author={Siebert, Lee and Simkin, Tom and Kimberly, Paul},
