### 2021-11-01 4d9cc3c
Create a separate function to print volcano data
In this commit, we have created the function: output_volcano_data(),
which takes a volcano name and prints the corresponding `ID profile`
(that is, the volcano data available for each of the different
volcanological criteria that PyVOLCANS uses to calculate single-
and multi-criteria volcano analogies).

For now, the printing of the ID profile only happens if the verbose
flag has been set to TRUE.

Re-arranging of the volcano data into a Pandas df that can be output
into a csv file will be implemented during subsequent commits.
