+++ pyvolcans/pyvolcans_func.py
@@ -449,18 +449,41 @@ def calculate_weighted_analogy_matrix(my_volcano, weights,
     volcans_result['ASt'] = \
         weighted_eruption_style_analogy[volcano_idx, ]
 
-    # TEST: PRINT VOLCANO DATA
+    return volcans_result
+
+
+def output_volcano_data(my_volcano, to_file=None, filename=None):
+    """
+    Prepares data for a given volcano (my_volcano) to be written either to the
+    standard output or to a comma-separated-value (csv) file.
+
+    Parameters
+    ----------
+    my_volcano : str or int
+        Target volcano selected by the user, as volcano name or volcano number
+    to_file : str, optional
+        Keyword argument that indicates whether the volcano data have to be
+        written onto a csv file (to_file='csv') or not (Default, to_file=None)
+    filename : str, optional
+        Keyword argument that indicates the filename to use for the optional
+        csv file containing the volcano data. Default filename=None.
+        Note that, if 'filename' is specified but 'to_file=None', the variable
+        'filename' is not used.
+
+    Returns
+    -------
+    """
+
+    # get the index for my_volcano
+    volcano_idx = convert_to_idx(my_volcano)
+
+    # details on variables of volcano data
     my_criteria_details = ['', 'Rock type proportions [F-P-T-Y-Z-X-B-A-D-R]',
                            '', 'P(VEI | eruption)[≤2-3-4-5-6-7-8]',
                            'Percentage of eruptions with phenomena'
                            ' [LF-BT-PH-WSF-TSU-PDC-DST-CF]']
-    my_volcano_country_test = \
-        volcans_result['country'].iloc[volcano_idx]
-    my_volcano_vnum_test = \
-        volcans_result['smithsonian_id'].iloc[volcano_idx]
-    print(f'ID profile for {my_volcano}, {my_volcano_country_test} '
-          f'({my_volcano_vnum_test}):')
 
+    # print volcano data
     for (criterion, detail) in zip(VOLCANO_DATA.keys(), my_criteria_details):
         criterion_data = VOLCANO_DATA[criterion]
         my_volcano_criterion_data = criterion_data[volcano_idx]
@@ -470,8 +493,9 @@ def calculate_weighted_analogy_matrix(my_volcano, weights,
             else:
                 print(f'{criterion}: {detail}: {my_volcano_criterion_data}\n')
 
-
-    return volcans_result
+#    if to_file == 'csv':
+#        result.to_csv(filename, sep=',', float_format='%.5f',
+#                      header=True, index=False, columns=my_columns)
 
 
 def get_analogies(my_volcano, volcans_result, count=10):
