### 2021-11-01 06369f5
Include details on tectonic-setting values in output_volcano_data()
A preliminary approach to do this has been implemented using a
dictionary of tectonic settings and their associated numeric values
used in PyVOLCANS. Better solutions to the issue may be explored in
subsequent commits.
