### 2021-11-02 14bc84c
Create a function to output the top-analogues ID profiles into csv
In this commit, a new function called: `output_many_volcanoes_data()`
has been incorporated. In it, the function: `output_volcano_data()`
is called iteratively, to obtain the ID profiles of all top analogue
volcanoes selected by the user (using `--count`).

These ID profiles are then written into a csv file. This functionality
of PyVOLCANS will be made optional via a designated flag (to be set up
in subsequent commits).
