### 2021-11-02 9864e4d
Add function to format volcano names into output filenames
This addition has simplified a few procedures in `pyvolcans.py`,
as well as improving the output filenames used for the 'a priori'
plots (.png) created in functions: `plot_bar_apriori_analogues()`,
and `plot_bar_better_analogues()`.
