### 2021-11-04 296a61f
Add test for format_volcano_name()
