### 2021-11-04 fd8e13f
Add data argument in output_volcano_data() to help testing
