### 2021-11-04 6b5bded
Add volcano names argument in output_volcano_data() to help testing
