### 2021-11-04 44e52de
Add test for output_volcano_data() function
