### 2021-11-04 BritishGeologicalSurvey/pyvolcans#16 @PTierz
Dear @meghanrjones,

I sincerely hope that this message finds you well.

I just wanted to let you know that we have preliminarily addressed Issue #16 in a new branch called [`volcano-id-profile`](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/volcano-id-profile).

We still need to add a few improvements before we merge the branch into [`main`](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/main) but please feel free to explore the new functionalities of the code and let us know if they appear to be satisfactory as regards your request.

We will aim at completing the merge by 15th November 2021.

Very many thanks in advance,
Pablo
