### 2021-11-12 970378c
Add test for Hekla source data
