+++ test/test_pyvolcans_func.py
@@ -3,6 +3,54 @@ import pytest
 from pyvolcans import pyvolcans_func
 
 
+def test_get_volcano_source_data():
+    # Arrange
+    expected = {
+        'name': 'Hekla',
+        'country': 'Iceland',
+        'smithsonian_id': 372070,
+        'tectonic_setting': 0.0,
+        'geochemistry': {
+            'Foidite': 0.,
+            'Phonolite': 0.,
+            'Trachyte': 0.,
+            'Trachyandesite/Basaltic trachyandesite': 0.,
+            'Phono-tephrite/Tephri-phonolite': 0.,
+            'Tephrite/Basanite/Trachybasalt': 0.,
+            'Basalt': 0.25,
+            'Andesite': 0.25,
+            'Dacite': 0.25,
+            'Rhyolite': 0.25,
+        },
+        'morphology': 0.39474,
+        'eruption_size': {
+            'VEI ≤2': 0.26667,
+            'VEI 3': 0.35,
+            'VEI 4': 0.28333,
+            'VEI 5': 0.1,
+            'VEI 6': 0.,
+            'VEI 7': 0.,
+            'VEI 8': 0.,
+        },
+        'eruption_style': {
+            'Lava flow and/or fountaining': 0.87692,
+            'Ballistics and tephra': 0.69231,
+            'Phreatic and phreatomagmatic activity': 0.,
+            'Water-sediment flows': 0.15385,
+            'Tsunamis': 0.01538,
+            'Pyroclastic density currents': 0.09231,
+            'Edifice collapse/destruction': 0.,
+            'Caldera formation': 0.,
+        }
+    }
+
+    # Action
+    source_data = pyvolcans_func.get_volcano_source_data('Hekla')
+
+    # Assert
+    assert source_data == expected
+
+
 @pytest.mark.parametrize('volcano, weights, expected', [
     ('Korath Range',
      {'tectonic_setting': 0.2,
