### 2021-11-12 8ec3bdd
Add first draft volcano_source_data function
