### 2021-11-12 BritishGeologicalSurvey/pyvolcans#16 @volcan01010
To pretty print a dictionary:

```python
import json

my_volcano = {'name': 'Hekla', 'other_attribute': 1}
print(json.dumps(my_volcano, indent=2, sort_keys=False))
```

`sort_keys` will keep them in the order that they were entered.
