### 2021-11-12 2b22d41
Complete volcano_source_data function
