+++ test/test_pyvolcans_func.py
@@ -9,7 +9,9 @@ def test_get_volcano_source_data():
         'name': 'Hekla',
         'country': 'Iceland',
         'smithsonian_id': 372070,
-        'tectonic_setting': 0.0,
+        'tectonic_setting': {
+                0.0: 'Rift Oceanic Crust'
+        },
         'geochemistry': {
             'Foidite': 0.,
             'Phonolite': 0.,
@@ -22,23 +24,23 @@ def test_get_volcano_source_data():
             'Dacite': 0.25,
             'Rhyolite': 0.25,
         },
-        'morphology': 0.39474,
+        'morphology': 0.39473684210526316,
         'eruption_size': {
-            'VEI ≤2': 0.26667,
+            'VEI ≤ 2': 0.26666666666666666,
             'VEI 3': 0.35,
-            'VEI 4': 0.28333,
+            'VEI 4': 0.2833333333333333,
             'VEI 5': 0.1,
             'VEI 6': 0.,
             'VEI 7': 0.,
             'VEI 8': 0.,
         },
         'eruption_style': {
-            'Lava flow and/or fountaining': 0.87692,
-            'Ballistics and tephra': 0.69231,
+            'Lava flow and/or fountaining': 0.8769230769230769,
+            'Ballistics and tephra': 0.6923076923076923,
             'Phreatic and phreatomagmatic activity': 0.,
-            'Water-sediment flows': 0.15385,
-            'Tsunamis': 0.01538,
-            'Pyroclastic density currents': 0.09231,
+            'Water-sediment flows': 0.15384615384615385,
+            'Tsunamis': 0.015384615384615385,
+            'Pyroclastic density currents': 0.09230769230769231,
             'Edifice collapse/destruction': 0.,
             'Caldera formation': 0.,
         }
