+++ pyvolcans/pyvolcans.py
@@ -30,6 +30,7 @@ from pyvolcans.pyvolcans_func import (
     convert_to_idx,
     plot_bar_apriori_analogues,
     plot_bar_better_analogues,
+    get_volcano_source_data,
     output_volcano_data,
     output_many_volcanoes_data,
     format_volcano_name,
@@ -146,6 +147,8 @@ def cli():
         if args.verbose:
             print(f'\nID profile for {volcano_name}, {my_volcano_country} '
                   f'({my_volcano_vnum}):')
+            id_profile = get_volcano_source_data(volcano_name)
+            print(id_profile)
             if args.output_volcano_data:
                 volcano_name_joined = format_volcano_name(volcano_name)
                 output_filename = Path.cwd() / \
