### 2021-11-12 57b4734
Print ID profile of target volcano using json
