### 2021-11-12 a9a436d
Adapt output_many_volcanoes_data() to call get_volcano_source_data() and output a json-format file
