### 2021-11-12 3fb8091
Remove unused output_volcano_data() function from the code
