+++ pyvolcans/pyvolcans_func.py
@@ -556,97 +556,6 @@ def get_volcano_source_data(my_volcano, data = VOLCANO_DATA,
     return result
 
 
-def output_volcano_data(my_volcano, data=VOLCANO_DATA, names=VOLCANO_NAMES,
-                        to_file=None, filename=None):
-    """
-    Prepares data for a given volcano (my_volcano) to be written either to the
-    standard output or to a comma-separated-value (csv) file.
-
-    Parameters
-    ----------
-    my_volcano : str or int
-        Volcano identifier, as volcano name or volcano number, for the volcanic
-        system from which the ID profile (volcano data) will be output
-    to_file : str, optional
-        Keyword argument that indicates whether the volcano data have to be
-        written onto a csv file (to_file='csv') or not (Default, to_file=None)
-    filename : str, optional
-        Keyword argument that indicates the filename to use for the optional
-        csv file containing the volcano data. Default filename=None.
-        Note that, if 'filename' is specified but 'to_file=None', the variable
-        'filename' is not used
-
-    Returns
-    -------
-    all_volcano_data : Pandas dataframe
-        Dataframe containing the data for all the volcanological criteria
-        (i.e. the ID profile) for the volcano used to run the function
-    """
-
-    # get the index for my_volcano
-    volcano_idx = convert_to_idx(my_volcano)
-
-    # details on variables of volcano data
-    my_criteria_details = ['', 'Rock type proportions [F-P-T-Y-Z-X-B-A-D-R]',
-                           '', 'P(VEI | eruption)[≤2-3-4-5-6-7-8]',
-                           'Percentage of eruptions with phenomena'
-                           ' [LF-BT-PH-WSF-TSU-PDC-DST-CF]']
-
-    # tectonic setting information
-    my_tectonic_setting_dict = {'Rift Oceanic Crust': 0,
-                                'Intraplate Oceanic Crust': 0.125,
-                                'Rift Intermediate Crust': 0.250,
-                                'Intraplate Intermediate Crust': 0.375,
-                                'Rift Continental Crust': 0.500,
-                                'Intraplate Continental Crust': 0.625,
-                                'Subduction Zone Oceanic Crust': 0.750,
-                                'Subduction Zone Intermediate Crust': 0.875,
-                                'Subduction Zone Continental Crust': 1.000}
-
-    # create a list to append volcano data to
-    volcano_data_list = []
-
-    # create a Pandas df with the basic volcano info
-    volcano_info=names.iloc[volcano_idx]
-    volcano_info.index = ['name', 'country', 'smithsonian_id']
-
-    # print volcano data
-    for (criterion, detail) in zip(data.keys(), my_criteria_details):
-        criterion_data = data[criterion]
-        my_volcano_criterion_data = criterion_data[volcano_idx]
-        volcano_data_list.append(
-                [np.around(np.array(my_volcano_criterion_data), 5)])
-        with np.printoptions(precision=3, suppress=True):
-            if criterion == 'tectonic_setting':
-                my_key_list = list(my_tectonic_setting_dict.keys())
-                my_val_list = list(my_tectonic_setting_dict.values())
-                position = my_val_list.index(my_volcano_criterion_data)
-                my_tectonic_setting_key = my_key_list[position]
-                print(f'{criterion}: {detail}: '
-                      f'{my_volcano_criterion_data:.3f} '
-                      f'({my_tectonic_setting_key})\n')
-            elif criterion == 'morphology':
-                print(f'{criterion}: {detail}: '
-                      f'{my_volcano_criterion_data:.3f}\n')
-            else:
-                print(f'{criterion}: {detail}: {my_volcano_criterion_data}\n')
-
[…]
