### 2021-11-12 c402c52
Remove momentary deactivation of printing
