### 2021-11-12 166ae79
Specify that verbose mode needs to be on to output the ID profiles into json-format files
