### 2021-11-12 ad3eb18
Move ID profile print to bottom and minor fix on analogues ID profile output
