### 2021-11-12 06dc4b7
Remove currently unneeded test for output_volcano_data() function
