### 2021-11-12 529c2e4
Change less-or-equal-than symbol to 'leq' string in dict keys
