### 2021-11-12 caa7a04
Add documentation for get_volcano_source_data() and remove ames as input argument
