### 2021-11-12 #3649 @PTierz
Dear @meghanrjones, @elbeejay and @jifarquharson,

I hope this message finds you well.

I would like to kindly ask you to have a look at [Merge Request #17](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/17) in our PyVOLCANS repository, and let us know whether you consider that the incorporated changes address [Issue #16](https://github.com/BritishGeologicalSurvey/pyvolcans/issues/16) satisfactorily.

A few changes could still be needed as @volcan01010 and/or @mobiuscreek double-check the merge request, but I think the majority of the work to address the aforementioned Issue has been incorporated into the merge request.

Looking forward to hearing from you regarding these new additions to PyVOLCANS. I sincerely hope that they can help the case of making our manuscript suitable for publication in the Journal of Open Source Software.

Very many thanks for all your work to improve our software.

Best regards,
Pablo
