### 2021-11-15 #3649 @elbeejay
Hi @PTierz,

Thank you for your efforts to add functionality in this new merge request.

You previously noted that additional tests would be written:

> Test coverage: we can certainly increase the test coverage and include the CLI component. For now, only some of the trickiest functions have tests

do you expect this work effort to take place in a separate branch / PR?

And I think a more minor but valid suggestion raised by @meghanrjones was:

> One small comment regarding the title/summary: since there is no API documentation for the Python functions, the fact that users interact with PyVOLCANS via the CLI is essential information for those interested in using the tool. Therefore, this information should be included in the title and/or summary.

I agree that at minimum this should be incorporated in the summary section of the paper, if not in the title itself to make it clear to readers/users that they will be interacting with the program via the command-line as opposed to from Python itself.

Thanks,
Jay
