### 2021-11-19 b8bd189
Add initial tests for cli functionality
