### 2021-11-19 9baf919
Fix a typo on Torfajokull volcano name
