### 2021-11-19 13fec86
Modify (slightly) the way of generating the json file with `-oad` flag
These changes circumvent the JSONDecodeError but there is still
an index error in the assertion part of the test.
