### 2021-11-22 #3649 @elbeejay
> Hi @elbeejay,
>
> Very many thanks for your message.
>
> I think we may add the tests for `pyvolcans.py` in a separate merge request from [the one concerning `volcano-id-profile` branch](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/17). Many thanks.
>
> Regarding the specification of the fact that the users interact with PyVOLCANS via the CLI, we had included that information in the [“Statement of Need” section of our manuscript](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/JOSS_paper/paper/paper.md), where we describe our package from a general to more detailed perspectives. Users can also quickly see that they will interact with PyVOLCANS via the command line in the [README.md file](https://github.com/BritishGeologicalSurvey/pyvolcans#readme), which they encounter when accessing the [PyVOLCANS repository](https://github.com/BritishGeologicalSurvey/pyvolcans). We sincerely hope that these specifications can be satisfactory enough. We would really like to keep the title of the manuscript as it is now: we do not see any inaccurate message in it and we think it is very important that it remains as simple and attractive as possible. Very many thanks in advance for your understanding.
>
> Please also note that we have now added the necessary acknowledgements to @jifarquharson and @meghanrjones, as reviewers of the manuscript, and to yourself, as Associate Editor, in the latest version of our manuscript (please see [commit 0b3e1e7 ](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/0b3e1e7c40ed1ec36a79b505d5c7e8faf0d1d01a)). Thank you very much for all your work and support.
>
> Best regards, Pablo (on behalf of the co-authors)

Thanks Pablo.

In the interest of our reviewers time, could you ping us all once PR#17 as well as the anticipated PR regarding the additional tests for `pyvolcans.py` have been merged? I think it will make more sense for us to all have another look once you and your collaborators are finished making, reviewing, and merging the various changes.

-Jay
