### 2021-11-23 #3649 @PTierz
Good morning @elbeejay (@jifarquharson, and @meghanrjones),

Many thanks for your message. Certainly. We will do so as soon as the two merges have been carried out. We expect this to happen by close of play on Friday 26th November.

Hopefully, the changes will be satisfactory for you and the reviewers. I personally think that the code will have significantly improved by the time we have incorporated the two PRs.

Very many thanks again for your suggestions to make `PyVOLCANS` a better product.

All the best,
Pablo
