### 2021-11-24 1ba1e46
Add test parameterisations for volcanological-criteria flags
Given the odd behaviour under `--verbose` mode (identified by
@mobiuscreek), where half the output is routed to stderr and
half to stdout, I have included both `out`, `err` in the assertion
part of `test_pyvolcans_output()`. We may want to fix this during
forthcoming commits. Many thanks.
