### 2021-11-24 afcb72e
Fix faulty test assertion (stdout, stderr)
The volcano_name used in the test parameterisations has also been
changed, and an additional test case for the flag `--count` has been
added. Finally, test name `test_write_files()` has been modified to
`test_write_json_files()`.
