### 2021-11-24 8d30636
Modify `test_write_json_files()` to make it pass
Given that `get_volcano_source_data()`, linked with `-ovd`, returns
a dict, and `output_many_volcanoes_data()`, linked with `-oad` returns
a list of dicts, the test was adapted to either check for the volcano
name of the target volcano (`-ovd`) or that of the top 1 analogue to
the target volcano (`-oad`). A few extra test parameterisations will
be added in the following commits.
