### 2021-11-24 734fbfd
Add extra test parameterisations for `test_write_json_files()`
