### 2021-11-24 67a2006
Add test for writing csv files
