### 2021-11-24 d9adc7e
Add test for weight errors in cli
