### 2021-11-25 3a3efd4
Expand test parameterisations for PyvolcansErrors
