+++ test/test_pyvolcans_func.py
@@ -1,6 +1,233 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri May 15 12:49:55 2020
+
+@author: Vyron Christodoulou, John A. Stevenson, Pablo Tierz
+         (British Geological Survey, The Lyell Centre,
+         Edinburgh, UK).
+"""
 import pytest
 
-from pyvolcans import pyvolcans_func
+import numpy as np
+import pandas as pd
+from pandas.testing import assert_frame_equal
+from pyvolcans.pyvolcans_func import (
+    fuzzy_matching,
+    match_name,
+    format_volcano_name,
+    get_volcano_idx_from_name,
+    get_volcano_name_from_idx,
+    get_volcano_number_from_name,
+    get_volcano_idx_from_number,
+    get_many_analogy_percentiles,
+    get_volcano_source_data,
+    calculate_weighted_analogy_matrix,
+    open_gvp_website,
+    plot_bar_apriori_analogues,
+    plot_bar_better_analogues,
+    set_weights_from_args,
+    VOLCANO_NAMES,
+    PyvolcansError
+)
+
+# pylint: disable=missing-docstring
+
+
+def test_volcano_idx():
+    idx = get_volcano_idx_from_name(volcano_name='Fuego')
+    assert idx == 1071
+
+
+def test_volcano_name():
+    name = get_volcano_name_from_idx(volcano_idx=1071)
+    assert name == 'Fuego'
+
+
+def test_fuzzy_matching():
+    volc_matches = fuzzy_matching('West Eifel')
+    assert 'West Eifel Volcanic Field' in volc_matches
+
+    # Test with limit
+    volc_matches_limit = fuzzy_matching('West Eiffel', limit=3)
+    assert 'West Eifel Volcanic Field' in volc_matches_limit
+
+
+def test_volcano_number():
+    number = get_volcano_number_from_name('Santorini')
+    assert number == 212040
+
+
+def test_format_volcano_name():
+    name = format_volcano_name('Tolima, Nevado del')
+    assert name == 'Tolima_Nevado_del'
+
+
+def test_set_weights_from_args_error():
+    args_dict = {'tectonic_setting': 99.00001,
+                 'geochemistry': 99.00001,
+                 'morphology': 99.00001,
+                 'eruption_size': 99.00001,
+                 'eruption_style': 99.00001}
+
+    with pytest.raises(PyvolcansError) as exc_info:
+        set_weights_from_args(args_dict)
+
+    msg = ("Sum of weights (495.000050000) is different from 1!"
+           " Please revise your weighting scheme.")
+
+    assert str(exc_info.value) == msg
+
+
+@pytest.mark.parametrize("args_dict, expected_weights", [
+    ({'tectonic_setting': 0.2,
+      'morphology': 0.2,
+      'eruption_size': 0.2,
+      'geochemistry': 0.2,
+      'eruption_style': 0.2}, [0.2, 0.2, 0.2, 0.2, 0.2]),
+    ({'tectonic_setting': None,
+      'morphology': None,
+      'eruption_size': None,
+      'geochemistry': None,
+      'eruption_style': None}, [0.2, 0.2, 0.2, 0.2, 0.2]),
+    ({'tectonic_setting': 1,
+      'morphology': None,
+      'eruption_size': None,
+      'geochemistry': None,
+      'eruption_style': None}, [1, 0, 0, 0, 0]),
+    ({'tectonic_setting': 0.2,
+      'morphology': 0.4,
+      'eruption_size': None,
+      'geochemistry': 0.2,
+      'eruption_style': 0.2}, [0.2, 0.4, 0, 0.2, 0.2])
+    ])
+def test_set_weights_from_args(args_dict, expected_weights):
[…]
