+++ test/test_pyvolcans.py
@@ -1,229 +1,75 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Fri May 15 12:49:55 2020
-
-@author: Vyron Christodoulou, John A. Stevenson, Pablo Tierz
-         (British Geological Survey, The Lyell Centre,
-         Edinburgh, UK).
-"""
+import subprocess
 import pytest
-
-import numpy as np
+import json
 import pandas as pd
-from pandas.testing import assert_frame_equal
-from pyvolcans.pyvolcans_func import (
-    fuzzy_matching,
-    match_name,
-    format_volcano_name,
-    get_volcano_idx_from_name,
-    get_volcano_name_from_idx,
-    get_volcano_number_from_name,
-    get_volcano_idx_from_number,
-    get_many_analogy_percentiles,
-    calculate_weighted_analogy_matrix,
-    open_gvp_website,
-    plot_bar_apriori_analogues,
-    plot_bar_better_analogues,
-    set_weights_from_args,
-    VOLCANO_NAMES,
-    PyvolcansError
-)
+from pathlib import Path
 
-# pylint: disable=missing-docstring
-
-
-def test_volcano_idx():
-    idx = get_volcano_idx_from_name(volcano_name='Fuego')
-    assert idx == 1071
+@pytest.mark.parametrize("input_args,expected",
+        [("--help", "usage: pyvolcans"),
+         ("-h", "usage: pyvolcans"),
+         ("--version", "v"),
+         ("-V", "v"),
+         # Odd behaviour with --verbose, half of the output is written to stderr and half to stdout
+         ("Hekla --verbose", "ID profile for Hekla, Iceland (372070):"),
+         ("Hekla -v", "ID profile for Hekla, Iceland (372070):"),
+         ("Hekla --apriori Santorini", "According to PyVOLCANS"),
+         ("Vesuvius -Ts 1 -v", "'tectonic_setting': 1.0"),
+         ("Vesuvius -G 1 -v", "'geochemistry': 1.0"),
+         ("Vesuvius -M 1 -v", "'morphology': 1.0"),
+         ("Vesuvius -Sz 1 -v", "'eruption_size': 1.0"),
+         ("Vesuvius -St 1 -v", "'eruption_style': 1.0"),
+         ("Fuego --count 50", "Top 50")])
+def test_pyvolcans_output(input_args, expected, capfd):
+    subprocess.run(['pyvolcans', *input_args.split()])
+    out, err = capfd.readouterr()
+    assert (expected in out or expected in err)
 
 
-def test_volcano_name():
-    name = get_volcano_name_from_idx(volcano_idx=1071)
-    assert name == 'Fuego'
+@pytest.mark.parametrize("input_args,expected",
+                         [("Hekla -ovd", "Hekla"),
+                          ("Hekla -oad", "Torfajokull"),
+                          ("Fuego -ovd", "Fuego"),
+                          ("Fuego -oad", "Klyuchevskoy"),
+                          ("Sabancaya -oad", "Peuet Sague")])
+def test_write_json_files(input_args, expected, tmp_path):
+    subprocess.run(['pyvolcans', '--verbose', *input_args.split()],
+                    cwd=tmp_path)
+    fname = list(Path(tmp_path).glob("*.json"))[0]
+    with open(fname) as f:
+        data = json.load(f)
+    #`-ovd` returns a dict, `-oad` returns a list of dict
+    if type(data) == dict:
+        assert data['name'] == expected
+    elif type(data) == list:
+        top1_analogue = data[0]
+        assert top1_analogue['name'] == expected
 
 
-def test_fuzzy_matching():
-    volc_matches = fuzzy_matching('West Eifel')
[…]
