### 2021-11-25 888cff5
Simplify input arguments for 	est_plot_bar*() functions
