### 2021-11-25 802c9db
Add a test for get_analogies() func
