### 2021-11-26 6f8576f
Use pip install via setup.py to install pyvolcans
