+++ .github/workflows/ci.yml
@@ -19,8 +19,7 @@ jobs:
        
       - name: Install dependencies
         run: |
-             pip install -r requirements.txt
-             pip install pytest
+          python -m pip install -e .[dev]
      
       - name: Run tests
         run: python -m pytest -vs
