### 2021-11-26 f8e8197
Align requirements.txt and setup.py
