### 2021-11-26 2a05155
Skip version number test in GitHub actions
