### 2021-11-26 56812b9
Remove unused exception handling
