### 2021-11-26 61d37e0
Add test for perfect analogues warning
