### 2021-11-26 BritishGeologicalSurvey/pyvolcans#16 @PTierz
Dear @meghanrjones,

I hope this message finds you well.

First of all, please let us thank you for your review of our `PyVOLCANS` manuscript and for the very useful suggestions, including raising this issue, which have helped us to significantly improve the software.

Regarding the current issue, please see the changes in the now merged `volcano-id-profile` branch (#17).

In the current version of the manuscript, running `pyvolcans Kilauea -v` should provide not only the Top 10 analogues to Kīlauea (using a default, equal-weighting scheme) but also the _ID profile_ for the target volcano, as stored in data from GVP 4.6.7 and the merged volcano morphology database used in [Tierz et al. (2019)](https://doi.org/10.1007/s00445-019-1336-3).

Finally, running `pyvolcans Kilauea -v --output_volcano_data` should additionally save a `.json` file containing Kīlauea's  _ID profile_, and running `pyvolcans Kilauea -v --output_analogues_data` should save a different `.json` file containing the _ID profile_ for all the Top 10 analogues to Kīlauea, the latter considering the default, equal-weighting scheme used.

We sincerely hope that these changes will satisfy your request. If this is the case, please feel free to close the issue yourself, or let me know that the changes look appropriate, so I can close the issue.

Thank you very much in advance for your support.

Best regards,
Pablo
