### 2021-11-26 #3649 @PTierz
Dear @meghanrjones, @jifarquharson and @elbeejay,

I sincerely hope this message finds you well.

Please let me thank you once again for all your work, feedback and suggestions on how to improve the `PyVOLCANS` code and our submission of it to the Journal of Open Source Software.

I am elated to let you know that the latest requested changes have now been incorporated into a new [`v1.3.0`](https://github.com/BritishGeologicalSurvey/pyvolcans/releases/tag/v1.3.0) version of the software. This new version includes two major groups of changes: (1) the added functionality of `PyVOLCANS` to output the _ID profiles_ (i.e. volcanological data available) for both the target volcano and any set of analogue volcanoes generated (please see [merged branch `volcano-id-profile`](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/17); and (2) a comprehensive set of new tests implemented to cover the CLI component in the [`pyvolcans.py` file](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/main/pyvolcans/pyvolcans.py) of the package (please see [merged branch `test-cli`](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/18)).

You can check the first major addition by running commands like: `pyvolcans Kilauea -v` (which prints the Top 10 analogues to Kīlauea as well as the _ID profile_ for the target volcano into the stdout); `pyvolcans Kilauea -v --output_volcano_data` (which saves a `.json` file containing Kīlauea's  _ID profile_); and `pyvolcans Kilauea -v --output_analogues_data` (which saves a different `.json` file containing the _ID profile_ for all the Top 10 analogues to Kīlauea, the latter considering the default, equal-weighting scheme used).

In terms of test coverage, you can now check that `pytest --cov=pyvolcans test\` should inform that over 85% and 90% of [`pyvolcans.py`](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/main/pyvolcans/pyvolcans.py) and [`pyvolcans_func.py`](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/main/pyvolcans/pyvolcans_func.py), respectively, are now covered by tests included in our `PyVOLCANS` package.

We most sincerely hope that these additions to our software can significantly help the case of it meriting publication on the Journal of Open Source Software.

Please do not hesitate to let us know if you required anything else from us.

Best regards,
Pablo (on behalf of the co-authors)
