### 2021-11-26 cbc8a5d
Add clarification to help text of --plot_apriori flag
