+++ pyvolcans/pyvolcans.py
@@ -283,13 +283,16 @@ def parse_args():
                         help=("Print debug-level logging output, ID profile "
                               " for the selected target volcano, and include "
                               "single-criterion analogy values, besides the "
-                              "total analogy values, in the PyVOLCANS results")
+                              "total analogy values, in the PyVOLCANS results.")
                         )
     parser.add_argument("-pa", "--plot_apriori", action="store_true",
-                        help=("Generate bar plots displaying the values of "
-                        "single-criterion and total analogy between the target"
-                        " volcano and any 'a priori' analogues chosen by the"
-                        " user.")
+                        help=("Generate bar plots displaying: (1) values of "
+                              "single-criterion and total analogy between the "
+                              "target volcano and any 'a priori' analogues "
+                              "chosen by the user; and (2) percentages of "
+                              "'better analogues' (for the target volcano) "
+                              "than each of the 'a priori' analogues, "
+                              "considering all volcanoes in the GVP database.")
                         )
     parser.add_argument("-S", "--save_figures", action="store_true",
                         help="Save all generated figures")
