### 2021-11-26 24069f7
Update README.md with latest `--plot_apriori` help text
