### 2021-11-26 #3649 @PTierz
Dear @meghanrjones, @jifarquharson and @elbeejay,

I just wanted to let you know that we have also updated the main [README file](https://github.com/BritishGeologicalSurvey/pyvolcans#readme) in the PyVOLCANS GitHub repository, as well as the [PyVOLCANS Wiki pages](https://github.com/BritishGeologicalSurvey/pyvolcans/wiki) to reflect all changes incorporated up to [PyVOLCANS v1.3.0](https://github.com/BritishGeologicalSurvey/pyvolcans/releases/tag/v1.3.0). If you had any comments or requests, please let us know.

Looking forward to hearing from you.
Thank you very much in advance for your time.

Best regards,
Pablo (on behalf of the co-authors)
