### 2021-11-29 BritishGeologicalSurvey/pyvolcans#16 @maxrjones
Thanks @PTierz. I will close this issue now and review the changes using the checklist in the joss-reviews issue.
