### 2021-12-07 04095ff
Add Papale (2021) and Marzocchi et al. (2021) references
