+++ paper/paper.bib
@@ -94,3 +94,25 @@ school={University of Edinburgh}
   publisher = {Venzke, E (ed.). Smithsonian Institution. Downloaded 09 Jun 2021. \url{https://doi.org/10.5479/si.GVP.VOTW4-2013}}
 }
 
+@article{Marzocchi:2021,
+  title={A unified probabilistic framework for volcanic hazard and eruption forecasting},
+  author={Marzocchi, Warner and Selva, Jacopo and Jordan, Thomas H},
+  journal={Natural Hazards and Earth System Sciences},
+  volume={21},
+  number={11},
+  pages={3509--3517},
+  year={2021},
+  publisher={Copernicus GmbH},
+  url={https://doi.org/10.5194/nhess-21-3509-2021}
+}
+
+@incollection{Papale:2021,
+  title={Some relevant issues in volcanic hazard forecasts and management of volcanic crisis},
+  author={Papale, Paolo},
+  booktitle={Forecasting and Planning for Volcanic Hazards, Risks, and Disasters},
+  pages={1--24},
+  year={2021},
+  publisher={Elsevier},
+  url={https://doi.org/10.1016/B978-0-12-818082-2.00001-9}
+}
+
