### 2021-12-07 #3649 @PTierz
Dear @meghanrjones, @jifarquharson and @elbeejay,

I hope this message finds you well.

I just wanted to inform you that I have added two recent, but very relevant, citations to our JOSS manuscript, which further show the future potential and applicability of VOLCANS/`PyVOLCANS` to the field of volcanic hazard assessment (please see commits [04095ff](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/04095fffe725ead13f3721236491804158e2db9c#diff-382635c9325bf3273d195ff1b8a44e5b11afd7d97addeb8863ea35feb98c1a07), [ed88cb6](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/ed88cb632eadf0bdfa433e795a1bd86bb2abe617#diff-382635c9325bf3273d195ff1b8a44e5b11afd7d97addeb8863ea35feb98c1a07)).

I considered that it would be important for future users of the code, and hopefully for future readers of our manuscript, to know more about the context and application of the presented software tool.

Thank you very much in advance.

Kind regards,
Pablo
