### 2021-12-08 #3649 @elbeejay
@PTierz just so you are aware, I will be asking you to merge the JOSS_paper branch into the main branch at the end of this process. I haven't checked if there are any merge conflicts but wanted to mention this.
