### 2021-12-15 #3649 @elbeejay
Hello @meghanrjones and @jifarquharson,

Wanted to reach out to you both about this review.

@meghanrjones - have you had a chance to look over the recent revision to the repository? No problem if not; please just update us with an anticipated timeline.

@jifarquharson - several revisions to the submission have taken place since your review was conducted. Would you be willing to have a second look through these changes? Please give us an estimate as to when you will be able to do this check-up.

Thank you both,
Jay
