### 2021-12-16 #3649 @jifarquharson
Hi @elbeejay,

> several revisions to the submission have taken place since your review was conducted. Would you be willing to have a second look through these changes? Please give us an estimate as to when you will be able to do this check-up.

I have been keeping up with these as they come in, and all the changes seem to be useful and proportional. I will re-test `pyVOLCANS` tomorrow and note any additional comments that might be warranted.

Regards,
Jamie
