### 2021-12-17 #3649 @jifarquharson
Dear @elbeejay and @PTierz et al.,

I have re-read the article, "PyVOLCANS: A Python package to flexibly explore similarities and differences between volcanic systems," and have tested the additional `pyVOLCANS` functionality (now v1.3.0). In particular, the improved
`--output_analogues_data` and `--output_volcano_data` options increase the range of applications of the tool. I can foresee immediate use-cases of this updated functionality. That being said, a small issue exists that hinders their full utility (there is no disambiguation of .json filenames, and files can be overwritten without warning).

Pending a fix to [#20](https://github.com/BritishGeologicalSurvey/pyvolcans/issues/20), I am really happy with the revised version and heartily recommend its publication. The minor amendments to the text are warranted, with the added references further demonstrating the timely nature of this contribution. I'm satisfied that all of the checklist items have been satisfied.

- [ ] Close [#20](https://github.com/BritishGeologicalSurvey/pyvolcans/issues/20).
