### 2021-12-17 cca2fd6
Disambiguate output filenames from PyVOLCANS by adding volcano_number
In previous versions, calling PyVOLCANS using target volcanoes with
exactly the same name (e.g. Sumbing, Santa Isabel) would generate
output files (via the `format_volcano_name()` function) that would
have identical filenames. Hence, issues related to overwriting the
output files for different target volcanoes existed (many thanks
@jifarquharson for [pointing it
out!](https://github.com/BritishGeologicalSurvey/pyvolcans/issues/20))

With the changes in this commit, the `format_volcano_name()` function
includes the volcano nu
[…]
