### 2021-12-17 e3422d1
Update 	est_format_volcano_name() to reflect the inclusion of the volcano_number as argument
