### 2021-12-17 BritishGeologicalSurvey/pyvolcans#20 @PTierz
Hi @jifarquharson,

Very many thanks for all your work checking the functionalities of `PyVOLCANS` and for raising this Issue, which is very relevant, especially for iterative runs of the program, as you were commenting. Many thanks.

We have addressed the Issue in commits e3422d1, cca2fd6, and the changes are incorporated in version [`v1.3.1`](https://github.com/BritishGeologicalSurvey/pyvolcans/releases/tag/v1.3.1) of the code. One advantage of updating the `format_volcano_name()` function (by adding the volcano_number as input argument) is that not only `.json` filenames can be made unique but also the `.csv` and `.png` files that may be generated using `PyVOLCANS`

If you, @meghanrjones, @mobiuscreek or @volcan01010 could double-check the acceptance criteria below and close the Issue if they were satisfied, it would be excellent.

Thank you very much once again for your time and effort.

Best regards,
Pablo

### Acceptance criteria
- [ ] `pyvolcans 261180 -v -ovd` generates a file called `Sumbing_261180_IDprofile.json` with the ID profile for Sumbing (261180).
- [ ] `pyvolcans 263220 -v -ovd` generates a file called `Sumbing_263220_IDprofile.json` with the ID profile for Sumbing (263220).
- [ ] `pyvolcans 353060 -v -oad` generates a file called `Azul_Cerro_353060_top10analogues_Ts0200G0200M0200Sz0200St0200_IDprofiles.json` containing the ID profiles for the top 10 analogues to Cerro Azul (353060), using an equal-weighting scheme.
- [ ] `pyvolcans 357060 -v -oad` generates a file called `Azul_Cerro_357060_top10analogues_Ts0200G0200M0200Sz0200St0200_IDprofiles.json` containing the ID profiles for the top 10 analogues to Cerro Azul (357060), using an equal-weighting scheme.
- [ ] `pyvolcans Fuego -w` generates a `.csv` file called `Fuego_342090_top10analogues_Ts0200G0200M0200Sz0200St0200.csv`, which contains the top 10 analogues to Fuego, using an equal-weighting scheme.
- [ ] `pyvolcans Sinabung --apriori Unzendake "Soufriere Hills" "Huila, Nevado del" "Guagua Pichincha" Redoubt -pa --save_figures` generates `.png` files named `Sinabung_261080_apriori_analogues_Ts0200G0200M0200Sz0200St0200.png` and `Sinabung_261080_better_analogues_Ts0200G0200M0200Sz0200St0200.png` containing the relevant plots for the _a priori analogues_ chosen.
