### 2021-12-17 #3649 @PTierz
Dear @jifarquharson,

Thank you very much for all the time and work that you have dedicated to the review of our `PyVOLCANS` code and manuscript. It is hugely appreciated.

Likewise, very many thanks for your support and positive comments about the tool. We sincerely hope that it can be used by and useful for as many volcano scientists and enthusiasts as possible.

We have just posted a comment on Issue [#20](https://github.com/BritishGeologicalSurvey/pyvolcans/issues/20), explaining how we have addressed it and suggesting some acceptance criteria to be checked before the Issue can be closed.

Thank you so much once again.

Best wishes,
Pablo (on behalf of the co-authors)
