### 2021-12-17 BritishGeologicalSurvey/pyvolcans#20 @jifarquharson
Hi @PTierz,

I can confirm that, with  [cca2fd6](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/cca2fd68e6cb79b55799b9f089e155983820672d), each of those commands works as expected. Thanks for making these changes. Happy to close this issue now.
