### 2021-12-17 #3649 @jifarquharson
Happy to report that with [`cca2fd6`](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/cca2fd68e6cb79b55799b9f089e155983820672d) and the update to [`v1.3.1`](https://github.com/BritishGeologicalSurvey/pyvolcans/releases/tag/v1.3.1), the disambiguation/overwriting problem ([#20](https://github.com/BritishGeologicalSurvey/pyvolcans/issues/20)) is fixed.
