### 2021-12-19 #3649 @meghanrjones
Dear all,

I also re-read the article “PyVOLCANS: A Python package to flexibly explore similarities and differences between volcanic systems” and tested out the functionality added since my initial review using version 1.3.1. I am pleased to report that the changes have satisfactorily addressed my concerns for the remaining checklist items "Substantial scholarly effort” and "Automated tests”. I have edited my review checklist accordingly and recommend publication in JOSS after removing the bold formatting on sections that represent changes in the article text. I appreciate the authors' efforts in making the results from VOLCANS accessible for scholarly research through an open-source tool.

I agree with the authors that the current form of the package will be a useful resource for instructors and researchers. While my recommendation for publication does not require additional changes, I advise that the authors consider adding badges to their readme to facilitate quick evaluation of the build/test status, test coverage, and PyPI project page. I also expect that the package would reach a broader audience if the authors add documentation examples for loading the .JSON and .csv output into commonly used Python objects such as a pandas.DataFrame or provide a complementary, simple Python API.

Sincerely,

Meghan Jones
