### 2021-12-20 #3649 @elbeejay
Thanks for the quick replies @jifarquharson and @meghanrjones, your comments and review of this work are greatly appreciated.

@PTierz I would like to ask that you finalize the `paper.md` file at this time, as Meghan pointed out there is bold formatting in sections that represent changes that are likely not intended to be bold in the final version. I would also suggest the authors consider the suggestions made by Meghan in the above post, namely:

 > consider adding badges to their readme to facilitate quick evaluation of the build/test status, test coverage, and PyPI project page

and

> add documentation examples for loading the .JSON and .csv output into commonly used Python objects such as a pandas.DataFrame or provide a complementary, simple Python API

While neither of these things is a requirement, I do agree that they would increase the user-friendliness of the package and may help people newer to Python or PyVOLCANS.

In any case, please notify me when you've finalized the paper so that I can go through and type-check it for any minor formatting or reference problems.

Thanks,
Jay
