### 2021-12-20 97cbb5b
Remove bold text from paper.md
