### 2021-12-20 #3649 @PTierz
Dear @meghanrjones and @elbeejay,

Thank you so very much for your messages and for all the work that you have dedicated to reviewing our `PyVOLCANS` software and manuscript. This has resulted in significant improvements to `PyVOLCANS` and we are greatly thankful for that.

Many thanks for your additional suggestions too. We will commit to start working on them after the Christmas break.

I have now removed the bold text from the [`paper.md`](https://github.com/BritishGeologicalSurvey/pyvolcans/blob/JOSS_paper/paper/paper.md) file (please see commit [`97cbb5b`](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/97cbb5b6f42a1a4d9cd140e8517e5350a7f9b663)).

Please do not hesitate to let me know about any other task that I should complete.

Best wishes,
Pablo (on behalf of the co-authors)
