### 2021-12-20 BritishGeologicalSurvey/pyvolcans#21 @volcan01010
Add status badges to GitHub README
### Description

Following a suggestion from @meghanrjones, adding status badges to the README will "facilitate quick evaluation of the build/test status, test coverage, and PyPI project page."

See https://github.com/openjournals/joss-reviews/issues/3649#issuecomment-997304926 for details.

### Acceptance criteria

GitHub landing page README contains:

+ [ ] Build status badge
+ [ ] Test coverage badge
+ [ ] PyPI version badge ([![PyPI version](https://badge.fury.io/py/pyvolcans.svg)](https://badge.fury.io/py/pyvolcans))
