### 2021-12-20 BritishGeologicalSurvey/pyvolcans#22 @volcan01010
Document loading output to Pandas dataframe
### Description

See https://github.com/openjournals/joss-reviews/issues/3649#issuecomment-997304926 for details.

### Acceptance criteria

+ [ ] Extra section in Wiki for "Further processing of pyVOLCANS output"
+ [ ] Sample code for reading analogue score CSV files into Pandas
+ [ ] Sample code for reading volcano profile JSON into Pandas (use `pd.json_normalise()` to flatten structure)
