### 2021-12-22 3e9397b
bib formatting
update formatting for the "Volcanoes of the World" citation
