+++ paper/paper.bib
@@ -91,7 +91,8 @@ school={University of Edinburgh}
   author = {{Global Volcanism Program}},
   title = {{Volcanoes of the World, v. 4.10.0 (14 May 2021)}},
   year = {2013},
-  publisher = {Venzke, E (ed.). Smithsonian Institution. Downloaded 09 Jun 2021. \url{https://doi.org/10.5479/si.GVP.VOTW4-2013}}
+  publisher = {Smithsonian Institution},
+  url={https://doi.org/10.5479/si.GVP.VOTW4-2013}
 }
 
 @article{Marzocchi:2021,
