### 2021-12-22 eb24e84
fix missing DOIs
update URL fields to DOI fields
