+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages={e2020GC009219},
   year={2020},
   publisher={Wiley Online Library},
-  url={https://doi.org/10.1029/2020GC009219}
+  doi={10.1029/2020GC009219}
 }
 
 @article{Tierz:2019,
@@ -19,7 +19,7 @@
   pages={76},
   year={2019},
   publisher={Springer},
-  url={https://doi.org/10.1007/s00445-019-1336-3}
+  doi={10.1007/s00445-019-1336-3}
 }
 
 @mastersthesis{White:2020,
@@ -34,7 +34,7 @@ school={University of Edinburgh}
   author={Loughlin, Susan C and Sparks, Steve and Brown, Sarah K and Vye-Brown, Charlotte and Jenkins, Susanna F},
   year={2015},
   publisher={Cambridge University Press},
-  url={https://doi.org/10.1017/CBO9781316276273}
+  doi={10.1017/CBO9781316276273}
 }
 
 @article{Newhall:2002,
@@ -46,7 +46,7 @@ school={University of Edinburgh}
   pages={3--20},
   year={2002},
   publisher={Springer},
-  url={https://doi.org/10.1007/s004450100173}
+  doi={10.1007/s004450100173}
 }
 
 @article{Newhall:2017,
@@ -57,7 +57,7 @@ school={University of Edinburgh}
   pages={184--199},
   year={2017},
   publisher={Elsevier},
-  url={https://doi.org/10.1016/j.jvolgeores.2017.08.003}
+  doi={10.1016/j.jvolgeores.2017.08.003}
 }
 
 @article{Cashman:2014,
@@ -68,7 +68,7 @@ school={University of Edinburgh}
   pages={28},
   year={2014},
   publisher={Frontiers},
-  url={https://doi.org/10.3389/feart.2014.00028}
+  doi={10.3389/feart.2014.00028}
 }
 
 @phdthesis{Simmons:2021,
@@ -76,7 +76,7 @@ school={University of Edinburgh}
   author={Isla Simmons},
   year={2021},
   school={University of Edinburgh},
-  url={https://doi.org/10.7488/era/1087}
+  doi={10.7488/era/1087}
 }
 
 @book{Siebert:2010,
@@ -92,7 +92,7 @@ school={University of Edinburgh}
   title = {{Volcanoes of the World, v. 4.10.0 (14 May 2021)}},
   year = {2013},
   publisher = {Smithsonian Institution},
-  url={https://doi.org/10.5479/si.GVP.VOTW4-2013}
+  doi={10.5479/si.GVP.VOTW4-2013}
 }
 
 @article{Marzocchi:2021,
@@ -104,7 +104,7 @@ school={University of Edinburgh}
   pages={3509--3517},
   year={2021},
   publisher={Copernicus GmbH},
-  url={https://doi.org/10.5194/nhess-21-3509-2021}
+  doi={10.5194/nhess-21-3509-2021}
 }
 
 @incollection{Papale:2021,
@@ -114,6 +114,6 @@ school={University of Edinburgh}
   pages={1--24},
   year={2021},
   publisher={Elsevier},
-  url={https://doi.org/10.1016/B978-0-12-818082-2.00001-9}
+  doi={10.1016/B978-0-12-818082-2.00001-9}
 }
 
