### 2021-12-22 #3649 @elbeejay
@PTierz and co-authors, I've opened up a [small PR](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/23) to modify a bit of formatting in the `.bib` file. I also made some suggestions for minor edits to the paper itself. The removal of non-DOI links in particular is something I think is important and will help the paper remain durable in the future. The PR should also help resolve some of the missing DOIs whedon has identified above.

Once you've taken action on the paper and the bibliography, please merge your paper branch into the main branch of the code.

Following that, there are a few items that need to be taken care of:

- [ ] Make a tagged release of the software and report the version tag for the archived code here
- [ ] Archive the reviewed code on Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] Check that the archived code (e.g., Zenodo) has the correct metadata. The title should match the paper title, the author list should be correct, and you can include authors' ORCID if you wish
- [ ] Provide the DOI of the archived code here

Thanks, please just reply with the above items once they are complete. Once that is done, I will briefly review the paper once more before referring it to one of the EiCs for publication.
