### 2021-12-23 41da41e
Re-format 'publisher field' in GVP (2013) reference
The reference now includes reference to editor (Ed Venzke) and download date.
