### 2021-12-23 6aff287
Rewording some sentences and updating references in paper.md
This follows the suggestions given in PR #23
