+++ paper/paper.md
@@ -44,7 +44,7 @@ Quantitative data to address these questions are scarce [@Loughlin:2015].
 While a handful of volcanoes (e.g. Etna, Italy; Kīlauea, USA; Merapi, Indonesia) have been extensively studied, hundreds of volcanic systems around the world remain poorly-understood.
 One possible mitigation to the issue of data scarcity in volcanology and volcanic hazard assessment is the use of _analogue volcanoes_ [@Newhall:2002; @Newhall:2017].
 These are volcanoes with similar characteristics to a data-scarce volcano of interest.
-Data and insights from the well-studied volcano(es) can be used to provide estimates for important variables, such as the number of eruptions during specific time windows or the size of those eruptions. Such methods have been used for many years but we have created the first tool to enable a structured and harmonised approach worldwide.
+Data and insights from the well-studied volcano(es) can be used to provide estimates for important variables, such as the number of eruptions during specific time windows or the size of those eruptions. Such methods have been used for many years, here we present the first tool enabling a structured and harmonised approach that can be applied worldwide.
 
 
 # Statement of need
@@ -52,10 +52,9 @@ Data and insights from the well-studied volcano(es) can be used to provide estim
 `PyVOLCANS` (Python VOLCano ANalogues Search) is an open-source tool that addresses the need for an objective, data-driven method for selection of analogue volcanoes.
 It is based on the results of VOLCANS [@Tierz:2019], a first-of-its-kind method to quantify the analogy (or similarity) between volcanic systems, based on a structured combination of five volcanological criteria: tectonic setting, rock geochemistry, volcano morphology, eruption size, and eruption style.
 `PyVOLCANS` provides a command-line interface to make the results from the VOLCANS study easily accessible to a wide audience.
-`PyVOLCANS` is a versatile tool for volcano scientists, with potential applications ranging from investigating commonalities between volcanic systems [@Cashman:2014] to supporting probabilistic volcanic hazard assessment at local, regional and global scales.
-It can also be used as a tool for [teaching and scientific outreach](https://twitter.com/Xeno_lith/status/1384416032526266369?s=20).
+`PyVOLCANS` is a versatile tool for volcano scientists, with potential applications ranging from investigating commonalities between volcanic systems [@Cashman:2014] to supporting probabilistic volcanic hazard assessment at local, regional and global scales. Exploring similarities and differences between volcanic systems using `PyVOLCANS` can also be useful for teaching and scientific outreach purposes.
 
-Users can easily derive data-driven sets of _top_ analogue volcanoes (i.e. those with highest analogy) to any volcanic system listed in the reference database for recent global volcanism: the [Volcanoes of the World Database](https://volcano.si.edu/list_volcano_holocene.cfm), hosted by the Global Volcanism Program of the Smithsonian Institution [@GVP:2013].
+Users can easily derive data-driven sets of _top_ analogue volcanoes (i.e. those with highest analogy) to any volcanic system listed in the reference database for recent global volcanism: the Volcanoes of the World Database, hosted by the Global Volcanism Program of the Smithsonian Institution [@GVP:2013].
 Users can also choose the number of _top_ analogue volcanoes to investigate and can customise the importance (i.e. weight) that is given to each of the five aforementioned volcanological criteria.
 Additionally, users can select a number of _a priori_ analogue volcanoes (i.e. volcanoes deemed as analogues by other means, such as expert knowledge) and assess their values of analogy with the target volcano to see how well they match on different criteria and if other volcanoes could be a better choice (\autoref{fig:figure1}).
 
@@ -64,7 +63,7 @@ ATs: Analogy in Tectonic setting; AG: Analogy in rock Geochemistry; AM: Analogy 
[…]
