### 2021-12-23 #3649 @PTierz
Hi @elbeejay,

Thank you so very much for all your help and support.

I proceeded with merging [your suggested PR](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/23) into the [`JOSS_paper` branch](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/JOSS_paper). I also added a couple of commits to the branch. I will wait for your confirmation that these changes look reasonable before I complete the [PR for the branch](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/19), OK?

Very many thanks in advance.

Best regards,
Pablo
