### 2021-12-24 #3649 @elbeejay
> Hi @elbeejay,
>
> Thank you so very much for all your help and support.
>
> I proceeded with merging [your suggested PR](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/23) into the [`JOSS_paper` branch](https://github.com/BritishGeologicalSurvey/pyvolcans/tree/JOSS_paper). I also added a couple of commits to the branch. I will wait for your confirmation that these changes look reasonable before I complete the [PR for the branch](https://github.com/BritishGeologicalSurvey/pyvolcans/pull/19), OK?
>
> Very many thanks in advance.
>
> Best regards, Pablo

@PTierz this looks good to me. At this time I'd say go ahead and merge the paper into the main branch and then proceed to complete the items on that checklist I provided above. If you notice something or want to make minor changes to the paper feel free to go ahead and do so, I'd just ask that you use the whedon commands (like those above) to regenerate the PDF and check the references to ensure nothing has broken in the paper or bibliography files.

Let me know if you have any questions as you are tagging or archiving the code.

Thanks,
Jay
