### 2021-12-29 #3649 @PTierz
Dear @elbeejay,

Thank you very much once again for all your support and assistance during the entire review process.

I have just merged the `JOSS_paper` branch into `main` and released version [`v1.3.2`](https://github.com/BritishGeologicalSurvey/pyvolcans/releases/tag/v1.3.2).

I have also archived [this version on Zenodo](https://zenodo.org/record/5809105#.Ycxiwdso85k) and obtained the following badge:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5809105.svg)](https://doi.org/10.5281/zenodo.5809105)

If there was anything else required from me, please do not hesitate to let me know.

Very many thanks for everything.

Best wishes,
Pablo
