### 2021-12-30 #3649 @elbeejay
Cool, everything looks good to me. Big thanks again to @meghanrjones and @jifarquharson for providing constructive and helpful reviews.

@PTierz and co-authors: At this point I will recommend PyVOLCANS for publication in JOSS. Next an Editor-in-Chief will review the submission and review thread here to double-check that we haven't missed anything and will either notify you of any minor changes needed or will publish your article in JOSS. Thanks for your cooperation, I hope you all found the process to be helpful.
