### 2021-12-31 #3649 @arfon
@jifarquharson, @meghanrjones – many thanks for your thoughtful and detailed reviews here and to @elbeejay for editing this submission! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
