### 2022-01-10 #3649 @elbeejay
@openjournals/dev is it possible to add a sentence to the acknowledgements section even though this has been technically published already? The sentence would read:

> Published with permission of the Executive Director of British Geological Survey (NERC-UKRI).

@PTierz please feel free to add further details as you deem appropriate.
