### 2022-01-20 #3649 @elbeejay
Pinging folks with the requested alteration test from @PTierz, just to make sure this doesn't get lost in the shuffle. Thanks all.

> Very many thanks for this, @elbeejay, and apologies for not having replied before.
>
> Dear @openjournals/dev, @openjournals/joss-eics,
>
> The only additions that I would like to make to the Acknowledgements are the following (in bold text):
>
> "_The research leading to these results has been mainly supported by the UK National Capability Funding (Innovation Flexible Fund programme) **and by the Global Geological Risk Platform of the British Geological Survey NC-ODA grant NE/R000069/1: Geoscience for Sustainable Futures** […] Moreover, we would like to sincerely thank a number of colleagues with whom we shared very insightful conversations about analogue volcanoes and/or PyVOLCANS: Chris Newhall, Isla Sim-
> mons, Adriano Pimentel, Julia Crummy, Gezahegn Yirgu, Charlotte Vye-Brown, Lara Smale,
> Karen Fontijn, Ben Clarke, Susanna Jenkins, Elly Tennant, Pierre Barbillon, Elaine Spiller,
> Philippa White, Teresa Ubide, Sebastián García, Victoria Olivera, Jeremy Pesicek, Vanesa
> Burgos Delgado, Einat Lev, Jonty Rougier, Willy Aspinall, Paolo Papale, Monse Cascante and
> Thomas Giachetti. **Published with permission of the Executive Director of British Geological Survey (NERC-UKRI).**_"
>
> Would it be possible to add those two sentences to the Acknowledgements without modifying the publishing date?
>
> I apologise in advance for having missed the inclusion of these sentences during the page-proofing stage.
>
> Very many thanks in advance.
>
> Kind regards,
> Pablo
