### 2022-01-23 #3649 @arfon
> @openjournals/dev is it possible to add a sentence to the acknowledgements section even though this has been technically published already? The sentence would read:

Yes, it's possible. @PTierz – please make the changes to your `paper.md` file (in the repo) and let me (@arfon) know when you've done this.
