### 2022-02-02 #3649 @PTierz
Dear @arfon,

Thank you very much for your message and your support.

Please accept my apologies for not having replied to this message considerably earlier. I had to be out of the office 22 Jan-1 Feb, and I am catching up with my emails only from this morning.

I have just added the small amending sentences to the Acknowledgements section via commit [1713702](https://github.com/BritishGeologicalSurvey/pyvolcans/commit/1713702886725e02edda4fb032a75d06a51831f9).

If there is anything else required from me, please do not hesitate to let me know. I sincerely hope that the publishing date/year (2021) can be kept, as this would be very important for me.

Very many thanks once again for your support.

Best wishes,
Pablo
