### 2022-02-03 #3649 @arfon
@PTierz – the paper is now updated. The new PDF might take ~24 hours to show as updated on the JOSS site.
