### 2022-02-03 #3649 @PTierz
Thank you very much for your support, @arfon, it is extremely appreciated.

I checked the [DOI link to the article](https://doi.org/10.21105/joss.03649) and it seems that it has been already updated!

Many thanks again.
Best wishes,
Pablo
